### 2020-09-11 #2666 @whedon
[PRE REVIEW]: Multiscale Solar Water Heating
**Submitting author:** @milicag (<a href="http://orcid.org/0000-0003-1640-3931">Milica Grahovac</a>)
**Repository:** <a href="https://github.com/LBNL-ETA/MSWH" target ="_blank">https://github.com/LBNL-ETA/MSWH</a>
**Version:** v1.0.0
