### 2020-09-11 #2666 @danielskatz
@milicag - where in this repo is the JOSS paper?  We also require an OSI-approved license, normally a file called LICENSE
