### 2020-09-11 #2666 @milicag
@danielskatz The paper is on a branch: https://github.com/LBNL-ETA/MSWH/tree/issue5_submit_to_joss. Let me know if you'd need it merged into the master and I can do that.

The license is in the l[egal.md](https://github.com/LBNL-ETA/MSWH/blob/issue5_submit_to_joss/legal.md) file.
