### 2020-09-11 #2666 @danielskatz
Which OSI-approved license is this?
