### 2020-09-11 #2666 @danielskatz
This is modified from https://opensource.org/licenses/BSD-3-Clause which we don't allow.

Can you break this off into a LICENSE file that is just the BSD license, without any changes other than in the first line?
