### 2020-09-11 #2666 @milicag
So you need me to place the https://github.com/LBNL-ETA/MSWH/blob/issue5_submit_to_joss/legal.md#license-agreement into a separate file?
