### 2020-09-11 #2666 @milicag
Can you tell me why the license must be in a separate file?
