### 2020-09-11 #2666 @milicag
[Here is the license file as a separate file.](https://github.com/LBNL-ETA/MSWH/blob/master/LICENSE.md)
