### 2020-09-11 #2666 @danielskatz
the problem is you are adding to the standard license, which we don't allow.  It also makes it harder to compare what you have with the standard license.

The open source community in general is good at dealing with standard files in standard locations, and this isn't standard.
