### 2020-09-11 #2666 @danielskatz
> Here is the license file as a separate file.

Thanks - but you have modified the 3rd clause from the standard
