### 2020-09-11 #2666 @milicag
Sure, no problem, I just like to understand. The modifications are how the license was granted by LBNL. What is not standard in the license? Thank you so much for any clarifications.
