### 2020-09-11 #2666 @milicag
I just did a detailed read and the license has the same 3 clauses and provides the same ability to the users.
