### 2020-09-11 #2666 @danielskatz
I'm sorry, now I realize that this is not a BSD 3-clause license but it is almost a Lawrence Berkeley National Labs BSD Variant License (BSD-3-Clause-LBNL) license: https://opensource.org/BSD-3-Clause-LBNL.

The only change that is needed is to remove the extra text at the start of the first line, so that this line starts with "Copyright".
