### 2020-09-11 #2666 @danielskatz
BSD-3-Clause-LBNL is an OSI-approved license, so once you make this small change, it will be fine.
