+++ LICENSE.md
@@ -1,7 +1,6 @@
 ## License Agreement
 
-Multiscale Solar Water Heating (MSWH) Copyright (c) 2019, The
-Regents of the University of California, through Lawrence Berkeley National
+Copyright (c) 2019, The Regents of the University of California, through Lawrence Berkeley National
 Laboratory (subject to receipt of any required approvals from the U.S.
 Dept. of Energy).  All rights reserved.
 
