+++ LICENSE.md
@@ -1,6 +1,7 @@
 ## License Agreement
 
-Copyright (c) 2019, The Regents of the University of California, through Lawrence Berkeley National
+Copyright (c) 2019 Multiscale Solar Water Heating (MSWH), The
+Regents of the University of California, through Lawrence Berkeley National
 Laboratory (subject to receipt of any required approvals from the U.S.
 Dept. of Energy).  All rights reserved.
 
