### 2020-09-11 #2666 @milicag
Just as fyi, if you were to write a parser for the license, LBNL first writes the name of the software, so as was written in my initial version:
```
Multiscale Solar Water Heating (SWH) Copyright (c) 2019,
```
I see how the https://opensource.org/BSD-3-Clause-LBNL switches the order, so I've aligned the license in the separate file with this one, so:
```
Copyright (c) 2019 Multiscale Solar Water Heating (MSWH)
```
