### 2020-09-11 #2666 @danielskatz
I'm sorry to be picky about this, but can you change the first line from
> Copyright (c) 2019 Multiscale Solar Water Heating (MSWH), The Regents of the University of California, through Lawrence Berkeley National Laboratory (subject to receipt of any required approvals from the U.S. Dept. of Energy). All rights reserved.

to
> Copyright (c) 2019, The Regents of the University of California, through Lawrence Berkeley National Laboratory (subject to receipt of any required approvals from the U.S. Dept. of Energy). All rights reserved.
