### 2020-09-11 d466608
updated license text to align with JOSS needs
