+++ LICENSE.md
@@ -1,8 +1,6 @@
-## License Agreement
+## License Agreement for Multiscale Solar Water Heating (MSWH)
 
-Copyright (c) 2019 Multiscale Solar Water Heating (MSWH), The
-Regents of the University of California, through Lawrence Berkeley National
-Laboratory (subject to receipt of any required approvals from the U.S.
+Copyright (c) 2019, The Regents of the University of California, through Lawrence Berkeley National Laboratory (subject to receipt of any required approvals from the U.S.
 Dept. of Energy).  All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
