### 2020-09-11 5cae480
updated article bibliography
