+++ paper.bib
@@ -5,7 +5,7 @@
 	pages = {928},
   publisher = {John Wiley & Sons, Inc., Hoboken, New Jersey},
   isbn = {978-0-470-87366-3},
-  year = 2013
+  year = {2013},
 }
 
 @misc{Srcc:2013,
@@ -104,21 +104,25 @@
 }
 
 @misc{Doecode:2019,
-title = {Multiscale Solar Water Heating (MSWH) v1.0},
-author = {Grahovac, Milica and Hosbach, Robert and Coughlin, Katie and Ganeshalingam, Mohan and Gerhart, Hannes},
-url = {https://doi.org/10.11578/dc.20190524.2},
-howpublished = {[Computer Software] \url{https://doi.org/10.11578/dc.20190524.2}},
-year = {2019},
-month = {may}
+	title = {Multiscale Solar Water Heating (MSWH) v1.0},
+	author = {Grahovac, Milica and Hosbach, Robert and Coughlin, Katie and Ganeshalingam, Mohan and Gerhart, Hannes},
+	url = {https://doi.org/10.11578/dc.20190524.2},
+	howpublished = {[Computer Software] \url{https://doi.org/10.11578/dc.20190524.2}},
+	year = {2019},
+	month = {may},
 }
 
 @article{Sparn:2014,
-	title = {Laboratory {Performance} {Evaluation} of {Residential} {Integrated} {Heat} {Pump} {Water} {Heaters}},
-	language = {en},
-	journal = {Renewable Energy},
-	author = {Sparn, B and Hudon, K and Christensen, D},
-	year = {2014},
-	pages = {77},
+	title = {Laboratory Performance Evaluation of Residential Integrated Heat Pump Water Heaters},
+	author = {Sparn, B. and Hudon, K. and Christensen, D.},
+	abstractNote = {This report discusses how a significant opportunity for energy savings is domestic hot water heating, where an emerging technology has recently arrived in the U.S. market: the residential integrated heat pump water heater. A laboratory evaluation is presented of the five integrated HPWHs available in the U.S. today.},
+	doi = {10.2172/1025650},
+	journal = {},
+	number = ,
+	volume = ,
+	place = {United States},
+	year = {2011},
+	month = {9},
 }
 
 @thesis{Gerhart:2019,
@@ -134,7 +138,7 @@ month = {may}
 	title = {Costs and {Benefits} of {Community} versus {Individual} {End}-use {Infrastructure} for {Solar} {Water} {Heating}},
 	institution = {California Energy Commission},
 	author = {Coughlin, Katie and Grahovac, Milica and Ganeshalingam, Mohan and Hosbach, Robert and Vossos, Vagelis},
-	year = {2020}
+	year = {2020, In Press}
 }
 
 @article{Grahovac:2020,
