+++ paper.bib
@@ -115,11 +115,8 @@
 @article{Sparn:2014,
 	title = {Laboratory Performance Evaluation of Residential Integrated Heat Pump Water Heaters},
 	author = {Sparn, B. and Hudon, K. and Christensen, D.},
-	abstractNote = {This report discusses how a significant opportunity for energy savings is domestic hot water heating, where an emerging technology has recently arrived in the U.S. market: the residential integrated heat pump water heater. A laboratory evaluation is presented of the five integrated HPWHs available in the U.S. today.},
 	doi = {10.2172/1025650},
-	journal = {},
-	number = ,
-	volume = ,
+	journal = {National Renewable Energy Lab. (NREL), Golden, CO},
 	place = {United States},
 	year = {2011},
 	month = {9},
