### 2020-09-16 #2666 @sjpfenninger
@brynpickering, @dalonsoa would you be able to review this submission for JOSS?
