### 2020-09-24 #2695 @whedon
[REVIEW]: Multiscale Solar Water Heating
**Submitting author:** @milicag (<a href="http://orcid.org/0000-0003-1640-3931">Milica Grahovac</a>)
**Repository:** <a href="https://github.com/LBNL-ETA/MSWH" target ="_blank">https://github.com/LBNL-ETA/MSWH</a>
**Version:** v2.0.0
