### 2020-09-24 #2695 @sjpfenninger
@milicag Can you confirm which branch of the repository the review should be looking at? I see paper.md both in `master` and `issue5_submit_to_joss`. And can you move paper.md and associated files into a subdirectory inside the repository root.
