### 2020-09-24 #2695 @milicag
@sjpfenninger Please review the master branch. I have a few clarity questions. Should the sub-directory be of a certain name? Is it correct to assume that the associated files are: `paper.md`, `paper.bib` and `LICENSE.md`?
