### 2020-09-24 #2695 @milicag
@sjpfenninger I moved the  `paper.md`, `paper.bib` and `LICENSE.md` into a new directory called `paper`. Let me know if you need anything else.
