### 2020-09-24 #2695 @sjpfenninger
@milicag Thanks - but `LICENSE.md` should always be in the root of your repository (this is independent of your JOSS submission)
