### 2020-09-24 #2695 @sjpfenninger
@openjournals/dev Any idea what is going wrong here? I suspect it is connected to the fact that `https://github.com/LBNL-ETA/MSWH/tree/issue5_submit_to_joss` is set as the submission repository.
