### 2020-09-24 #2695 @milicag
@sjpfenninger Thanks, I moved the license back to the root. Just as a note, the paper compiled fine when it was on the root of the master. I deleted that branch as the PR was merged.
