### 2020-10-13 LBNL-ETA/MSWH#8 @brynpickering
Exceptions in example notebooks
I just tried running the following notebooks from the `scripts` folder: 'MSWH System Tool.ipynb', 'MSWH System Tool - Playground.ipynb', and 'MSWH System Tool - for CEC Report.ipynb' and couldn't get them to run in their entirety, namely:

1. they depend on seaborn, which isn't in the mswh requirements
2. they raise an exception in the `Individual retrofit solar thermal system` section (~cell 55): ``ERROR:mswh.system.models:Check project output of the backup - timestep sum preformed in this method and the total are not equal.``
3. There's also another exception in  'MSWH System Tool - Playground.ipynb', cell 38: ``NameError: name 'cold_cz' is not defined``

Note: this issue comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/2695)
