### 2020-10-13 LBNL-ETA/MSWH#8 @milicag
Thanks @brynpickering! I will add seaborn to the `setup.py`

Was the `MSWH System Tool.ipynb` running fine? The playground does not need to be functional at all times, so I may move it to a different branch.
