### 2020-10-13 LBNL-ETA/MSWH#8 @milicag
Sure! Would you mind me asking what do you exactly mean with pin. Would it be sufficient to tag the repo and provide information about the tag in the notebooks?
