### 2020-10-13 82bcfee
added seaborn to setup.py, added items to .gitignore
