### 2020-10-15 7b1bb5e
added reasonable rounding, updated publications listed on the readme, started excluding irrelevan imports from notebooks
