+++ setup.py
@@ -41,7 +41,6 @@ setup(
         'numpy>=1.16.1'
         'future',
         'plotly>=3.2.0',
-        'seaborn',
         'nbformat>=5.0.4',
         'psutil>=5.4.7',
         'django>=2.1.7',
