### 2020-10-16 LBNL-ETA/MSWH#10 @brynpickering
Installation instructions cleanup
The README installation instructions are a bit confusing and occasionally miss some steps:

1. It's a bit difficult to read the conda/virtualenv instructions since you switch between them on different lines. Would a code block on installation with conda and another with virtualenv be possible? You can always put comments in those code blocks to describe what's going on.

2. You also need to add `pip` as a conda dependency. I expected it to be installed when installing python, but that seems to have changed.

3. A nice cross-platform method to install `git-lfs` is via conda forge, so you may find it easier to give instructions purely with conda:
```
conda create -n mswh -c conda-forge python=3.6 pip git-lfs
conda activate mswh
git lfs install
git clone ...
pip install -e .
```

4. `git lfs clone` is [deprecated](https://github.com/git-lfs/git-lfs/pull/2526) and they now recommend using `git clone`.

5. I generally recommend pointing people to [miniconda](https://docs.conda.io/en/latest/miniconda.html) rather than the full Anaconda distribution, since it has a much lower disk usage.

Note: this issue comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/2695)
