### 2020-10-16 LBNL-ETA/MSWH#11 @brynpickering
pandas >v1.0 causes test failure
On a fresh install of MSWH, using the README instructions, The `mswh.system.tests.test_models.SystemTests.test_solar_electric` test fails with the following error trace:

```python
Traceback (most recent call last):
  File "/path/to/MSWH/mswh/system/tests/test_models.py", line 834, in test_solar_electric
    self.r['p_el_res_use']]],
  File "/path/to/conda-env/lib/python3.6/site-packages/pandas/core/indexing.py", line 873, in __getitem__
    return self._getitem_tuple(key)
  File "/path/to/conda-env/lib/python3.6/site-packages/pandas/core/indexing.py", line 1055, in _getitem_tuple
    return self._getitem_tuple_same_dim(tup)
  File "/path/to/conda-env/lib/python3.6/site-packages/pandas/core/indexing.py", line 750, in _getitem_tuple_same_dim
    retval = getattr(retval, self.name)._getitem_axis(key, axis=i)
  File "/path/to/conda-env/lib/python3.6/site-packages/pandas/core/indexing.py", line 1099, in _getitem_axis
    return self._getitem_iterable(key, axis=axis)
  File "/path/to/conda-env/lib/python3.6/site-packages/pandas/core/indexing.py", line 1037, in _getitem_iterable
    keyarr, indexer = self._get_listlike_indexer(key, axis, raise_missing=False)
  File "/path/to/conda-env/lib/python3.6/site-packages/pandas/core/indexing.py", line 1254, in _get_listlike_indexer
    self._validate_read_indexer(keyarr, indexer, axis, raise_missing=raise_missing)
  File "/path/to/conda-env/lib/python3.6/site-packages/pandas/core/indexing.py", line 1316, in _validate_read_indexer
    "Passing list-likes to .loc or [] with any missing labels "
KeyError: "Passing list-likes to .loc or [] with any missing labels is no longer supported. The following labels were missing: Index(['Tank Balancing Error (Heat Rate)'], dtype='object'). See https://pandas.pydata.org/pandas-docs/stable/user_guide/indexing.html#deprecate-loc-reindex-listlike"
```

This requires either pinning pandas in your requirements to stay below version 1.0, or updating the code to handle >=v1.0. pandas v1.x has been out for a while now, so I would suggest pinning to >v1.x and updating code accordingly.

Note: this issue comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/2695)
