### 2020-10-16 LBNL-ETA/MSWH#11 @milicag
As we use this code in a few analyses and have the code profiled to a satisfactory speed, I would prefer to simply limit the pandas package version and leave the pandas update to a different time in case we use it for a new project. I hope this is not a problem on your end.
