### 2020-10-18 96145c0
removed obsolete label, checked code works with latest pandas and numpy versions
