### 2020-10-18 LBNL-ETA/MSWH#11 @milicag
This was actually an issue with a deprecated label, that stayed in @hannesb0 models as I did some updates after he was done with his Master Thesis research. I've tested with the latest `pandas` and `numpy` and no edits were needed, so no need to limit versions for now!
