### 2020-10-18 LBNL-ETA/MSWH#11 @milicag
In fact, I started using a more descriptive label instead:
```
            'q_ovrcool_tank': 'Tank Overcool (Heat Rate)',
```
As computers do not really know the 2nd law of thermodynamics and I am willing to `partially forgive it occasionally`, I use the above variable to perform the appropriate accounting and reporting.
