### 2020-10-18 LBNL-ETA/MSWH#10 @milicag
- [x] pip installation instruction
- [x] `git-lfs` update
- [x] improve `virtualenv` instruction formatting
- [x] recommend miniconda
