### 2020-10-18 4b3d0de
updated python version to 3.8, edited readme
