### 2020-10-18 LBNL-ETA/MSWH#13 @nmstreethran
JOSS Review: Installation instructions and Django application name
Hi @milicag, I am [reviewing your JOSS submission](https://github.com/openjournals/joss-reviews/issues/2695) and have some initial comments regarding the installation instructions in the README file, and the Django web application. I will open a new issue once I have fully explored the code and functionality. I will also address the software paper and documentation in separate issues. Let me know if you have any questions. Thanks.

## Setup and installation

### Creating a virtual environment

The link to the guide on virtual environments should be updated to <https://packaging.python.org/guides/installing-using-pip-and-virtual-environments/>.

Creating a new virtual environment on Linux should be done using the `venv` package (i.e., same as macOS; `virtualenv` is recommended for Python 2). From the link above:

> If you are using Python 3.3 or newer, the venv module is the preferred way to create and manage virtual environments. venv is included in the Python standard library and requires no additional installation.

I think the command for creating a virtual environment on Linux with a specific Python version could be removed, as it only works with `virtualenv`.

### `ipykernel`

`ipykernel` must be installed in the environment using `pip` or `conda` prior to running the command outlined in this step. From the [docs](https://ipython.readthedocs.io/en/stable/install/kernel_install.html):

> The Jupyter Notebook and other frontends automatically ensure that the IPython kernel is available. However, if you want to use a kernel with a different version of Python, or in a virtualenv or conda environment, you’ll need to install that manually.

Alternatively, `ipykernel` can be added to the dependencies list and installed after creating and activating the virtual environment to prevent this issue.

Users without admin privileges must include the `--user` flag when registering the kernel:

```sh
python -m ipykernel install --user --name mswh
```

### Other suggestions

- (line 8; paragraph 3) tankeless -> tankless
- (line 8; paragraph 4) link to the database has a typo; should be `mswh/comm/mswh_system_input.db`
- (line 11) latex -> LaTeX
- (line 11) html -> HTML
- (line 27) repo -> repository
- (lines 42, 46) MacOS / Mac -> macOS
- (line 66) jupyter -> Jupyter
- (lines 66, 86) python -> Python
- (line 88, sentence 3) can created -> can **be** created

## Django Web Framework Deployment

The web application is named 'Solar Water Heating' instead of 'Multiscale Solar Water Heating'. Is this intended?
