### 2020-10-18 LBNL-ETA/MSWH#14 @nmstreethran
JOSS Review: Authorship and software paper
[JOSS review](https://github.com/openjournals/joss-reviews/issues/2695): comments on authorship and the software paper

## Contribution and authorship

The review criteria asks to check that "the full list of paper authors seems appropriate and complete". However, the list of authors of the software paper (4) does not match the list of contributors in the 'About' section of the README file, `setup.py`, Sphinx HTML documentation, and the DOE CODE submission (5), or the Sphinx LaTeX documentation (6). Also, Hannes Gerhart appears twice in the LaTeX documentation authors list.

## Software paper

### A statement of need

This section clearly explains the problems the software will solve, but does not mention who the target audience is (e.g., homeowners, researchers).

### Quality of writing

The paper could use some work in terms of sentence structure and punctuation. For example, these sentences from line 55:

> The component models we either identified in the existing literature and created a custom Python implementation, or we developed new models. In our implementation we implemented the following existing or new models:

... could be rephrased as: *"The component models were either developed from scratch or implemented in Python based on existing models identified in literature. We implemented the following models:"*.

I can make more specific suggestions if you open a pull request regarding this issue, if you prefer.

Minor suggestions:

- (line 66) 1h -> 1 hour
- (line 72) CA -> California
- (line 81) DOE code -> DOE CODE
- (line 81) master thesis -> master's thesis
- (line 93) Modelica building library `@Wetter:2014` -> Modelica building library **by** `@Wetter:2014`
- (line 97) PC with 12 cores -> computer with a 12-core processor

### References

#### SRCC. (2013, January)

The URL for this reference is no longer available.

#### Blair, N., et al. (2014)

This seems to be a [NREL Technical Report](https://www.nrel.gov/docs/fy14osti/61019.pdf), but the bibliography says it is published in *Renewable Energy*.
