### 2020-10-18 LBNL-ETA/MSWH#15 @nmstreethran
JOSS Review: Documentation
[JOSS review](https://github.com/openjournals/joss-reviews/issues/2695): comments on the documentation

## A statement of need

A statement of need is required in the documentation -- same as the software paper.

## Installation instructions

The set-up and installation instructions should be in the documentation too, rather than directing users to view the README file on the GitHub repository.

## Functionality documentation

`scripts/Project Level MSWH System Tool.ipynb` and `scripts/Project Level SWH System Tool.ipynb` have been mentioned in `mswh/system/components.py` and `mswh/system/models.py`, but these files do not exist.

## Automated tests

Steps to run the functional and unit tests should be included here.

## Community guidelines

There should be a section with guidelines on how to contribute to the code, report issues, and seek support. It should mention that development takes place on GitHub. The README file says to check the documentation for guidance but I cannot seem to find any.

## Other suggestions

`docs/source/models.rst`

- (line 29) tankeless -> tankless
- (line 29) basecase -> base case
- (line 42) CSI -> California Solar Initiative

`mswh/comm/create_or_update_db.py`

- (line 140) monochrystalline -> monocrystalline
- (line 141) polychrystalline -> polycrystalline

`mswh/comm/label_map.py`

- (line 21) Occupancy

`mswh/comm/sql.py`

- (line 44) tables
- (line 160) execute

`mswh/system/components.py`

- (line 27) timeseries
- (line 929) aperture (?)
- (line 1248) medium
- (line 1497) temperature
- (line 1754) predefined temperature
- (line 2263) square
- (line 2964) Initial
- (line 2974) parallel
- (line 2975) circulation

`mswh/tools/plots.py`

- (line 286) length

`mswh/tools/unit_converters.py`

- (lines 47, 66) milion -> million
- (lines 56, 75) mili -> milli
- (lines 58, 77) mikro -> micro
