### 2020-10-18 LBNL-ETA/MSWH#14 @milicag
Thanks @nmstreethran!

Here are some clarifications and the boxes for things I intend to address in either the current or a new PR.

Authorship is correct according to the rules for the JOSS journal after discussion with the team. A person that contributed to a part of code that was removed in the course of time opted to be listed in acknowledgments instead as a co-author.

- [x] I'll make sure to have Hannes appear only once.

- [x] Add target audience to statement of need

I like to keep my style if it is easy to understand the meaning. I don't think you need to rewrite my paper, with all due respect. I appreciate you offer though and if something simply is not clear, please point to it and I am happy to rewrite.

- [x] Address all minor suggestions

- [x] Fix 2 references
