### 2020-10-18 LBNL-ETA/MSWH#13 @milicag
Thanks @nmstreethran! @brynpickering also asked for changes in the readme, so I intend to blend your suggestions in. I do like to insist on a specific python version, so I maybe I might need to keep the `virtualenv`? It is what I still use with python3.8, that is my main development python version at the moment. It is one thing what one finds online and the other what one finds useful as a practitioner. I can mention it as an alternative for folks who care less about the python version.

Doesn't `ipykernel` get installed with jupyter notebook or jupyter lab? If not, I can add it separately.

To do:

- [x] address other suggestions
- [x] django package name - update to MSWH
- [x] add Jupyter installation info
