### 2020-10-18 LBNL-ETA/MSWH#15 @milicag
@nmstreethran This is very helpful! Please see my comments below. The boxes represent what I intend to address, as in my responses in the rest of the issues you've created.

Where in the JOSS documentation do you require the code package to have documentation and describe the required form of that documentation? In my understanding JOSS only requires the paper to satisfy the JOSS paper requirements, which seems extremely logical.

I like to not repeat myself, so the installation will stay in the readme as it is fully accessible to those that read the sphinx docs, that are not really required by JOSS in the first place.

- [x] update notebook names to latest version in t `models.py` and `components.py`

The readme provides an info on [how to run tests under contributing](https://github.com/LBNL-ETA/MSWH#contributing). Would you like to have a subtitle that says `Automated tests` in addition to that? I could do that.

- [x] address `Other suggestions`. Thanks for doing this!
