### 2020-10-18 LBNL-ETA/MSWH#13 @milicag
@nmstreethran Please see my intended approach above and let me know if you have any objections. This applies to issues #15 and #14 as I have left similarly structured comments in those issues too. After I hear from you on all 3 issues I will start addressing them.
