### 2020-10-18 02d57d6
corrected journal category for one referece
