+++ paper/paper.bib
@@ -73,7 +73,7 @@
 @article{Blair:2014,
 	title = {System {Advisor} {Model}, {SAM} 2014.1.14: {General} {Description}},
 	language = {en},
-	journal = {Renewable Energy},
+	journal = {National Renewable Energy Lab. (NREL), Golden, CO},
 	author = {Blair, Nate and Dobos, Aron P and Freeman, Janine and Neises, Ty and Wagner, Michael and Ferguson, Tom and Gilman, Paul and Janzou, Steven},
 	year = {2014},
 	pages = {19},
