+++ paper/paper.bib
@@ -11,7 +11,7 @@
 @misc{Srcc:2013,
 	title = {{SRCC} {Standard} 100 - {Minimum} {Standards} {For} {Solar} {Thermal} {Collectors}},
   author = {SRCC},
-	url = {https://iccsafe.org/wp-content/uploads/is_stsc/SRCC_Standard_100_For_Solar_Collectors.pdf},
+	url = {https://solar-rating.org/programs/og-100-program/},
 	urldate = {2018-10-15},
 	publisher = {Solar Rating \& Certification Corporation},
 	month = jan,
