### 2020-10-18 LBNL-ETA/MSWH#13 @milicag
@hannesb0 May I leave it up to you to update the django application name? I don't know where I would need to change that. From my perspective it can also stay as is. <- never mind, I just did a string replace in the `web` folder.
