### 2020-10-18 7c85fea
fixed typos and capitalization #13
