### 2020-10-18 LBNL-ETA/MSWH#13 @milicag
This is now all addressed in PR #12.
