### 2020-10-18 68da48d
removed duplicate co-author listing #14
