### 2020-10-18 ac18224
fixed typos and capitalization in paper.md file #14
