### 2020-10-18 a530fab
updated example notebook name #15
