### 2020-10-18 LBNL-ETA/MSWH#16 @milicag
Release and final steps for JOSS publication
At appropriate points in the JOSS review:

- tag the codebase with release 2.0.0
- In `models.rst` add a sentence in the Sphinx documentation telling users that development primarily takes place on the GitHub repository, and that they can refer to the contributing guidelines and README for instructions on how to run tests
- make sure the links on the gh-pages work (also those from docstrings)
- regenerate gh-pages and the pdf of the documentation
- in the `paper.md` update the hyperlinks to point to release 2.0.0 files rather than master
- make sure to address https://joss.readthedocs.io/en/latest/submitting.html#the-review-process
- unless Hannes confirms his third affiliation may stay, remove it.

@nmstreethran @brynpickering
