### 2020-10-19 ece8fa9
fixed typos indicated in #15
