### 2020-10-19 LBNL-ETA/MSWH#15 @milicag
I addressed the issue in PR #12.
