### 2020-10-19 f222f5c
added subsection for automated testing #15
