### 2020-10-19 LBNL-ETA/MSWH#14 @milicag
@nmstreethran Thank you for raising all the issues! I addressed this in PR #12.
