### 2020-10-19 02a53bc
added jupyter installation info #13
