### 2020-10-19 LBNL-ETA/MSWH#13 @nmstreethran
@milicag, thanks for addressing the issues! The changes look fine to me, apart from a couple of things:

- The instructions for using `conda` could be moved to the top. Otherwise, `conda` users will go through two separate instructions for installing Git LFS.
- Since you're keeping `virtualenv`, wouldn't it be better to change the link to the virtual environment guide in README.md to `virtualenv`'s docs? The current link could confuse users as it says `venv` is preferred for newer Python versions.

> Doesn't `ipykernel` get installed with jupyter notebook or jupyter lab? If not, I can add it separately.

Yes, `ipykernel` is a dependency of `notebook`, which is a dependency of `jupyterlab`. As long as one of these packages is installed in the virtual environment prior to registering the kernel, the installation should proceed. The users following the `conda` instructions must install `notebook` or `jupyterlab` separately as well, as neither are included in the requirements in `setup.py`.

EDIT: I also suggest changing 'repo' to 'repository'.
