### 2020-10-19 LBNL-ETA/MSWH#15 @nmstreethran
Thanks @milicag. I am following the JOSS review checklist; each heading in this issue refers to a checklist item. You can see the checklist in https://github.com/openjournals/joss-reviews/issues/2695.

The software documentation is indeed part of your JOSS submission. From the [Review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html):

> There should be sufficient documentation for you, the reviewer to understand the core functionality of the software under review. A high-level overview of this documentation should be included in a `README` file (or equivalent).

So, the documentation should include the items I've mentioned, and the README file should contain an overview.

The other changes look good to me, apart from a couple of things:

- `docs/source/models.rst` (line 29) basecase -- it's written as 'base case' in README.md, and 'basecase' here
- `mswh/system/components.py` (line 1248) medium
