### 2020-10-19 LBNL-ETA/MSWH#14 @nmstreethran
Thanks for addressing the issues, @milicag. It was not my intention to rewrite your paper; it was just a suggestion and you do not need to go through with it. 'Quality of writing' is one of my review checklist items.

I do believe you need to edit line 55. I think it should be *'The component models were ...'*. I'll have to re-read the paper first and let you know afterwards if anything else is unclear.

~I don't see any changes made to the bibliography file regarding the two references. Everything else looks fine.~ Changes to the bibliography file have been committed directly to the main branch, so it's all good.
