### 2020-10-19 LBNL-ETA/MSWH#14 @milicag
yes, I felt bad having errors in the bibliography, sorry for not writing on that, there were many details!

I intend to go through and address the remaining issues tonight as needed.
