### 2020-10-19 LBNL-ETA/MSWH#10 @brynpickering
Thanks for making the changes @milicag. The instructions seem to have lengthened as a result, whereas I would say that minimising the length of installation instructions is key to being approachable to non-technical users.
If you have a non-technical user, they don't care about exactly what's happening (cloning, virtual environments, dependencies, pip -e flag, etc.), but most likely just want to get the web interface up and running on their machine.
If they are a technical user, then they will know how to install a pip package locally and will have a personal preference for conda/virtaulenv - i.e. there's no need to instruct them on the process.
Based on that, I would have a single code block in your README, and the more detailed instructions in the documentation. This would ensure you're not duplicating installation instructions (as noted in #15), whilst not polluting the README with excessive detail. All that requires then is to choose your preference of conda/virtualenv and have a single code block for it. For conda, it would be:

```
 conda create -n mswh -c conda-forge -c plotly python=3.8 pip git-lfs jupyter plotly-orca
 conda activate mswh
 git lfs install
 git clone https://github.com/LBNL-ETA/MSWH.git
 pip install -e .
```

Note that this installs the necessary elements to run a jupyter notebook (I include the whole jupyter package, since non-technical users will have no way to use `ipykernel` by itself) and to save plots with `orca`. Does it include redundant dependencies if someone doesn't use the notebook or plotting functionality in the end? Yes, but then the time to run and resulting memory footprint is low enough not to matter. The upside is that non-technical users will have everything without needing to deal with error messages and understanding how to install packages within their environment after the fact. Technical users will install via pip in whatever way they like, and then they won't get those redundant packages unless they want them at a later date (or can choose the more lightweight `ipykernel` for accessing MSWH from their jupyter notebook/lab session).

If a non-technical user loves it and wants to dig deeper, then your documentation will be their source, and right there will be your extended instructions along with explanation as to what's going on.
