### 2020-10-19 LBNL-ETA/MSWH#10 @milicag
I like that idea. @nmstreethran, do you find this to be a suitable solution? It satisfies my need to not repeat myself and it is going to make the instructions short and easy! All the good things.
