### 2020-10-19 LBNL-ETA/MSWH#17 @brynpickering
Improve code readability
OK, the title makes it seem like the code isn't readable - I actually find it to be quite clear and clean! However, there are a few things that can be a bit jarring. Here I've split it into three sub-issues:

1. The code could do with generally being formatted to improve readability to be a bit more [PEP 8](https://www.python.org/dev/peps/pep-0008/)ish, e.g. the code column is very narrow and unnecessary backslashes turn up in quite a few places. The easiest approach to not have to think about manual formatting is to just use [Black](https://github.com/psf/black), which will reformat your code to PEP 8 standard with a one-liner.

2. Something Black can't do is update variable names. In some instances these are a bit difficult to understand. There's no real reason to excessively shorten variable names in Python. A prime example is your lookup dictionaries for labels: `Converter.s`, `Converter.c`, and `Converter.r`; out of context they make no sense. Another example is `spec_hea_con` which is much easier to understand as `specific_heat_conductivity` (although perhaps this should be `thermal_conductivity`?). Full phrases, maybe with missing words are better than each word being shortened (i.e. `heat_conductivity` or `specific_heat` are both more informative than `spec_hea_con`, even if they're not perfectly "correct" descriptions in the thermodynamics community).

3. Finally, there's lots of repetition of filling dictionaries, which could either be simplified with helper functions or more `for` loops. As loops, e.g.
```python
self.params_piping[self.s['pipe_spec_hea_con']] =\
    params.loc[params[self.s['param']] == self.s['pipe_spec_hea_con'],
               self.s['param_value']].values[0]
self.params_piping[self.s['pipe_ins_thick']] = params.loc[
    params[self.s['param']] == self.s['pipe_ins_thick'],
    self.s['param_value']].values[0]
self.params_piping[self.s['dia_len_exp']] = params.loc[
    params[self.s['param']] == self.s['dia_len_exp'],
    self.s['param_value']].values[0]
```
could be
```python
for i in ['pipe_spec_hea_con', 'pipe_ins_thick', 'dia_len_exp']:
    self.params_piping[self.s[i]] = params.loc[params[self.s['param']] == self.s[i], self.s['param_value']].values[0]
```
Since these repetitive code snippets are generally based on DataFrames, you could also just map the index to the long names and then transform the result into a dictionary:
```python
df.index = df.index.map(self.s)
self.params_piping = df.to_dict()
```

Note 1: this issue comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/2695)
Note 2: from the perspective of the JOSS review, you can also stick with what you have, this is more of a side-note issue
