### 2020-10-19 LBNL-ETA/MSWH#17 @milicag
Thanks @brynpickering!

I find some of your suggestions useful and can address them in PR. However, please note that we currently do not have funding assigned to this project, so there will be a limited amount of improvement to already functional code that I will be able to do, with all my best efforts. I appreciate your understanding!
