### 2020-10-20 LBNL-ETA/MSWH#10 @nmstreethran
@milicag, yes, this solution is perfect!

EDIT: I think you'll need to install `notebook` (to install Jupyter Notebook) or `jupyterlab`, not `jupyter`.
