### 2020-10-22 LBNL-ETA/MSWH#10 @milicag
To self: For good info on git-lfs provide: https://github.com/git-lfs/git-lfs/wiki/Installation link.
