### 2020-10-22 LBNL-ETA/MSWH#18 @brynpickering
System cannot be generated or simulated on default 'load'
Trying to generate the project level system model entirely depending on defaults causes an Exception:

```python
[1] from mswh.system.models import System
[2] System()
[out] INFO:mswh.system.models:No weather data got passed to converters. Setting default scalar values for ambient temperature, 293.15, and solar irradiation, 800.
INFO:mswh.system.models:The system did not receive any load requirement. Setting hot water draw to 0.01 m3/h.
<ipython-input-32-3a8e98677fa5> in <module>()
----> 1 System()

/path/to/mswh/system/models.py in __init__(self, sys_params, backup_params, weather, sys_sizes, backup_sizes, loads, timestep, log_level)
    179
    180         self.timestep = timestep
--> 181         self.num_timesteps = len(self.load)
    182
    183     @property

TypeError: object of type 'float' has no len()
```
Would be fixed by setting the default load as an array, but I don't know how long you want that default array to be (one hour, i.e. `self.load = np.array([0.01])`?, one year of hourly data?)

If updating the code to `self.load = np.array([0.01])` on line 108 in `mswh.system.models`, the system model can be generated, but it then fails on not having any seasonal information. For instance, if I try to simulate it, I get:

```python
INFO:mswh.system.components:Storage parameters have not been passed to the class. Using default parameters.
sim---------------------------------------------------------------------------
<ipython-input-9-6b28b606d09a> in <module>()
----> 1 system_model.simulate()

/path/to/mswh/system/models.py in simulate(self, type)
   1093             elif type == 'gas_tank_wh':
   1094                 cons_total, sys_res = \
-> 1095                     self.conventional_gas_tank()
   1096                 proj_total = 0.
   1097                 sol_fra = 0.

/path/to/mswh/system/models.py in conventional_gas_tank(self)
   1015         self.cons_total.at[0, self.r['gas_use_w']] = (
   1016             res[self.r['gas_use']] *
-> 1017             self.it_is_winter[1:]).sum() * self.timestep
   1018
   1019         self.cons_total.at[0, self.r['gas_use_s']] = (

AttributeError: 'System' object has no attribute 'it_is_winter'
```

Note: this issue comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/2695)
