### 2020-10-22 LBNL-ETA/MSWH#18 @milicag
This is not the intended functionality that you are trying to get. I can add a message saying that.
