### 2020-10-22 LBNL-ETA/MSWH#19 @brynpickering
(git)ignore test results
The output PNGs of tests should be in the gitignore, so they don't pollute the repository. Easiest is probably to ignore all PNGs, but you might want to point it only on test directories, if you envision having images elsewhere in the repository.

Note: this issue comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/2695)
