### 2020-10-22 LBNL-ETA/MSWH#18 @brynpickering
I would go with removing defaults wherever you don't intend to have defaults. If you do that, I would also remove the code that attempts to deal with 'defaults' (e.g. `load=None`), since it isn't intended to work anyway.
