### 2020-10-22 LBNL-ETA/MSWH#20 @brynpickering
Misspellings in error logging message variables
Some things picked up by my linter:

`system.tests.test_models:L60`: 'inpath' not defined
`system.tests.test_source_and_sink:L41`: 'inpath' not defined and 'log' not initialised
`system.components:L744&810`: 'Q_dem' not defined
`system.components:L1045&1115`: 'mgs' is a misspelling
`system.source_and_sink:L747`: 'mg' is a misspelling
`system.source_and_sink:L772`: 'meg' is a misspelling
`system.tools.unit_converters`:  'log' not initialised

The error logs not being correctly set up also point to the invocation of these messages not being tested. That is, you don't know at the moment that these errors will be raised as expected. Adding tests to cover these would be good.

Note: this issue comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/2695)
