### 2020-10-22 LBNL-ETA/MSWH#19 @milicag
I commit the images to the repository, so I do need to track them with git. They are in the correct place on the repo.
