### 2020-10-22 LBNL-ETA/MSWH#19 @milicag
Yes, they get visually inspected each time I run the tests for any issues, so I commit them from time to time. If I ignore them then I won't be able to commit them when I'd like to. The motivation is that I do want that visual information on the repo, tests are for developers and developers should know how to deal with those and ignore when needed and push up should there be changes to the test. Non-technical user will not run tests, rather will look at primarily the notebooks, potentially spin up the web app.
