### 2020-10-22 LBNL-ETA/MSWH#18 @brynpickering
OK, sure. I think the least intrusive would be to remove the defaults in the function args, since their presence gave me the impression that I didn't need to define anything. If the code to handle loads as `None` is still there for some bespoke case, then requiring it to be explicitly set will stop anyone doing it 'accidentally'.
