### 2020-10-22 LBNL-ETA/MSWH#18 @milicag
Can find a way to prevent that. Thanks. The simulation does require weather data, loads, performance data, etc., so I will point the user who wishes to perform something similar to what you tried to the example notebook where things are set up as required.
