### 2020-10-22 LBNL-ETA/MSWH#17 @milicag
@brynpickering Short names are used for convenience in indexing. `spec_hea_con` is also super simple to read. The code snipped you suggest refactoring works and is readable. I don's see a real reason to introduce those types of edits.

I do find that applying Black is a good idea.
