### 2020-10-22 LBNL-ETA/MSWH#19 @milicag
With this I would like to close this issue if that is ok?
