### 2020-10-22 LBNL-ETA/MSWH#10 @milicag
@nmstreethran : I think @brynpickering suggests `jupyter` here:

```
conda create -n mswh -c conda-forge -c plotly python=3.8 pip git-lfs jupyter plotly-orca
```

I personally usually install `jupyter lab` these days. If we keep arguing about this level of detail we will not get a chance to show the actual benefit of having these simulation models in Python. If you both agree that this is something that is interesting for JOSS please make it possible for me to help that effort out, tidy up the essentials and ensure needed functionality. Otherwise it might become uninteresting for me if we keep re-arranging the installation instructions.
