### 2020-10-23 LBNL-ETA/MSWH#10 @nmstreethran
@milicag Alright, I'll leave the installation instructions as it is.
