### 2020-10-23 LBNL-ETA/MSWH#10 @brynpickering
Yeah, let's go with jupyterlab, since it's also my preference for personal use (and a newbie might as well start with the more feature-complete version of Jupyter), so:

```
conda create -n mswh -c conda-forge -c plotly python=3.8 pip git-lfs jupyterlab plotly-orca
```
And then just link to the [JupyterLab documentation](https://jupyterlab.readthedocs.io/en/latest/) for people to get up and running with that.
