### 2020-10-23 LBNL-ETA/MSWH#17 @brynpickering
This is the problem of being too used to your own code, since I had no idea what `spec_hea_con` was referring to. Given that I'm already familiar with terms used in thermodynamic analyses, it shows that the variable names are somewhat impenetrable.

This is not a problem for someone to use the finished product, but readability of the backend will be necessary if you hope for others to understand and contribute to the codebase.

Since this isn't central to the JOSS review requirements, I am not going to delay the review process for it! Just seemed worth pointing out, in case you wanted to spend time on it.
