### 2020-10-23 LBNL-ETA/MSWH#19 @brynpickering
OK, I'll close this. I'm not convinced that test results should be cluttering the core package, since everyone then has to download those images, including the majority who will never run any tests. Saying that, the package becomes quite large when accessing the databases, so it probably won't be noticed!
