### 2020-10-25 daf80fc
split installation instruction in simple and advanced sections
