### 2020-10-25 438b1db
edit to readme, and improvement to a sentence as recommended by reviewer #14
