### 2020-10-25 LBNL-ETA/MSWH#11 @milicag
Updated to python 3.8.
