### 2020-10-25 LBNL-ETA/MSWH#15 @milicag
Revised README and split the installation instruction into two sections - simple and advanced.
