### 2020-10-25 LBNL-ETA/MSWH#17 @milicag
The `docstrings`  can be used to read what each of the `kwargs` are, including `spec_hea_con`. I did find one method that uses it and does not list it under `Parameters`, so I will fix that while addressing this issue.
