### 2020-10-25 LBNL-ETA/MSWH#18 @milicag
The docstrings, in addition to the example notebooks, are providing examples on how to setup the inputs.
