+++ mswh/comm/create_or_update_db.py
@@ -24,7 +24,7 @@ log.setLevel(logging.DEBUG)
 # occupancies in the template db as well.
 # ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
-test_db_name = 'mswh_system_input.db'
+test_db_name = "mswh_system_input.db"
 
 # set write_new to True if you'd like to reinitiate the db. A copy of the
 # 'swh_system_input.db' will be saved with a 'bckp_' prefix.
@@ -35,30 +35,43 @@ sql_scripts = list()
 # ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 # scripts to drop tabes if rewriting them
 
-sql_scripts.append('''DROP TABLE IF EXISTS `sys_1_system_list`;
-''')
-sql_scripts.append('''DROP TABLE IF EXISTS `sys_2_system_configurations`;
-''')
-sql_scripts.append('''DROP TABLE IF EXISTS `sys_3_components`;
-''')
+sql_scripts.append(
+    """DROP TABLE IF EXISTS `sys_1_system_list`;
+"""
+)
+sql_scripts.append(
+    """DROP TABLE IF EXISTS `sys_2_system_configurations`;
+"""
+)
+sql_scripts.append(
+    """DROP TABLE IF EXISTS `sys_3_components`;
+"""
+)
 
 # component sizing tables
 
-sql_scripts.append('''DROP TABLE IF EXISTS `comp_1_sizing_regression`;
-''')
-sql_scripts.append('''DROP TABLE IF EXISTS `discrete_component_sizes`;
-''')
+sql_scripts.append(
+    """DROP TABLE IF EXISTS `comp_1_sizing_regression`;
+"""
+)
+sql_scripts.append(
+    """DROP TABLE IF EXISTS `discrete_component_sizes`;
+"""
+)
 
 # component performance parameters
 
-sql_scripts.append('''DROP TABLE IF EXISTS `component_performance`;
-''')
+sql_scripts.append(
+    """DROP TABLE IF EXISTS `component_performance`;
+"""
+)
 
 # ++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 # DB schema
 
 # lists system configurations
-sql_scripts.append('''CREATE TABLE `sys_1_system_list`
+sql_scripts.append(
+    """CREATE TABLE `sys_1_system_list`
 (
     `System ID` INTEGER NOT NULL,
     `System` TEXT NOT NULL,
@@ -66,9 +79,11 @@ sql_scripts.append('''CREATE TABLE `sys_1_system_list`
     `Retrofit` BOOLEAN NOT NULL,
 
     PRIMARY KEY (`System ID`)
-);''')
+);"""
+)
 
-sql_scripts.append('''CREATE TABLE `sys_3_components`
+sql_scripts.append(
+    """CREATE TABLE `sys_3_components`
 (
     `Component ID` INTEGER NOT NULL,
     `Component` TEXT NOT NULL,
@@ -76,10 +91,12 @@ sql_scripts.append('''CREATE TABLE `sys_3_components`
     `Component Size Unit` TEXT NOT NULL,
 
     PRIMARY KEY (`Component ID`)
-);''')
+);"""
+)
 
 # defines system configurations
-sql_scripts.append('''CREATE TABLE `sys_2_system_configurations`
+sql_scripts.append(
+    """CREATE TABLE `sys_2_system_configurations`
 (
     `System ID` INTEGER NOT NULL,
     `Component ID` INTEGER NOT NULL,
[…]
