+++ web/system/apps.py
@@ -3,5 +3,6 @@ from django.apps import AppConfig
 # See here for more information on how to use this file:
 # https://docs.djangoproject.com/en/2.1/ref/applications/
 
+
 class SystemConfig(AppConfig):
-    name = 'system'
+    name = "system"
