+++ mswh/system/tests/test_models.py
@@ -28,10 +28,10 @@ class SystemTests(unittest.TestCase):
     """Unit tests for the project level
     system models.
     """
+
     @classmethod
     def setUpClass(self):
-        """Assigns values to test variables.
-        """
+        """Assigns values to test variables."""
         random_state = np.random.RandomState(123)
 
         self.plot_results = True
@@ -47,7 +47,8 @@ class SystemTests(unittest.TestCase):
         # water draw profile
 
         weather_db_path = os.path.join(
-            os.getcwd(), 'mswh/comm/mswh_system_input.db')
+            os.getcwd(), "mswh/comm/mswh_system_input.db"
+        )
 
         db = Sql(weather_db_path)
 
@@ -56,16 +57,15 @@ class SystemTests(unittest.TestCase):
             # {table name : sheet name} form
             inputs = db.tables2dict(close=True)
         except:
-            msg = 'Failed to read input tables from {}.'
+            msg = "Failed to read input tables from {}."
             log.error(msg.format(inpath))
 
-        source_and_sink = SourceAndSink(
-            input_dfs=inputs)
+        source_and_sink = SourceAndSink(input_dfs=inputs)
 
         # SF climate is 03, 16 is cold
         self.weather = source_and_sink.irradiation_and_water_main(
-            '03',
-            method='isotropic diffuse')
+            "03", method="isotropic diffuse"
+        )
 
         # community scale household occupancies for 4 households
         occ_com = [4, 4, 3, 5]
@@ -74,83 +74,104 @@ class SystemTests(unittest.TestCase):
 
         # are people at home during the day in any of the households:
         # 'y' or 'n'
-        at_home_com = ['n', 'n', 'n', 'n']
-        at_home_ind = ['n']
+        at_home_com = ["n", "n", "n", "n"]
+        at_home_ind = ["n"]
 
         loads_com, peakload_com = SourceAndSink._make_example_loading_inputs(
             inputs,
             self.c,
             random_state,
             occupancy=occ_com,
-            at_home=at_home_com)
+            at_home=at_home_com,
+        )
 
-        loads_indiv, peakload_indiv = SourceAndSink._make_example_loading_inputs(
+        (
+            loads_indiv,
+            peakload_indiv,
+        ) = SourceAndSink._make_example_loading_inputs(
             inputs,
             self.c,
             random_state,
             occupancy=occ_ind,
-            at_home=at_home_ind)
+            at_home=at_home_ind,
+        )
 
         # scaled loads to match 300 L/day
         load_array_val = (
-            loads_indiv['End-Use Load'][0] *
-            (300 * 365 * .001) /
-            loads_indiv['End-Use Load'][0].sum())
+            loads_indiv["End-Use Load"][0]
+            * (300 * 365 * 0.001)
+            / loads_indiv["End-Use Load"][0].sum()
+        )
 
         loads_indiv_val = pd.DataFrame(
             data=[[1, occ_ind[0], load_array_val]],
-            columns=[self.c['id'], self.c['occ'], self.c['load_m3']])
+            columns=[self.c["id"], self.c["occ"], self.c["load_m3"]],
+        )
 
         # performance parameters
 
         # solar thermal
         sol_the_sys_params = pd.DataFrame(
-            data=[[self.s['the_sto'], self.s['f_upper_vol'], .5],
-                  [self.s['the_sto'], self.s['ins_thi'], .085],
[…]
