+++ web/system/migrations/0002_remove_configuration_result.py
@@ -6,12 +6,12 @@ from django.db import migrations
 class Migration(migrations.Migration):
 
     dependencies = [
-        ('system', '0001_initial'),
+        ("system", "0001_initial"),
     ]
 
     operations = [
         migrations.RemoveField(
-            model_name='configuration',
-            name='result',
+            model_name="configuration",
+            name="result",
         ),
     ]
