### 2020-10-25 LBNL-ETA/MSWH#18 @milicag
I implemented the following error message:
```
>>> System()
Please consult example setup either in the test suite or in the example notebooks to appropriatelly set up a System instance.
Traceback (most recent call last):
  File "<stdin>", line 1, in <module>
  File "/Users/mgrahovac/repos/MSWH/mswh/system/models.py", line 95, in __init__
    raise ValueError
ValueError
>>>
```
