### 2020-10-25 LBNL-ETA/MSWH#20 @milicag
- [x] system.tests.test_models:L60: 'inpath' not defined
- [x] system.tests.test_source_and_sink:L41: 'inpath' not defined and 'log' not initialised
- [x] system.components:L744&810: 'Q_dem' not defined
- [x] system.components:L1045&1115: 'mgs' is a misspelling
- [x] system.source_and_sink:L747: 'mg' is a misspelling
- [x] system.source_and_sink:L772: 'meg' is a misspelling
- [x] system.tools.unit_converters: 'log' not initialised
