### 2020-10-25 580669c
addressed part of the errors caught by the reviewer's linter
