### 2020-10-25 adfc708
addressed all #20
