+++ mswh/tools/unit_converters.py
@@ -1,6 +1,8 @@
 import numpy as np
 import pandas as pd
 
+log = logging.getLogger(__name__)
+
 
 class UnitConv(object):
     '''Unit conversions using conversion parameters from
