+++ mswh/system/tests/test_source_and_sink.py
@@ -9,7 +9,9 @@ from mswh.comm.sql import Sql
 from mswh.system.source_and_sink import SourceAndSink
 from mswh.comm.label_map import SwhLabels
 
-logging.basicConfig(level=logging.DEBUG)
+logging.basicConfig(level=logging.INFO)
+log = logging.getLogger(__name__)
+log.setLevel(logging.DEBUG)
 
 
 class SourceAndSinkTests(unittest.TestCase):
@@ -38,7 +40,7 @@ class SourceAndSinkTests(unittest.TestCase):
             inputs = db.tables2dict(close=True)
         except:
             msg = 'Failed to read input tables from {}.'
-            log.error(msg.format(inpath))
+            log.error(msg.format(weather_db_path))
 
         self.weather = SourceAndSink(
             input_dfs=inputs)
@@ -65,7 +67,7 @@ class SourceAndSinkTests(unittest.TestCase):
 
         # expected result
         cold_irrad_on_tilt = np.array(
-            [0., 0., 0., 0., 0., 0., 0., 3.62615245, 
+            [0., 0., 0., 0., 0., 0., 0., 3.62615245,
              349.78144662, 601.93687406, 823.2893731, 898.07605806,
              927.52023249, 825.36501459, 714.4711444, 463.85830573,
              0., 0., 0., 0., 0., 0., 0., 0.])
