+++ web/system/functions.py
@@ -10,6 +10,7 @@ from datetime import timedelta, datetime
 import numpy as np
 import pandas as pd
 import logging
+
 log = logging.getLogger(__name__)
 
 """
@@ -27,45 +28,49 @@ from mswh.comm.sql import Sql
 
 # Get labels associated with swh_sys project
 from mswh.comm.label_map import SwhLabels
+
 SYSTEM_LABELS = SwhLabels().set_prod_labels()
 CONSUMER_LABELS = SwhLabels().set_hous_labels()
 
 # Edit this dictionary to add new components
 COMPONENTS = {
-    'converter' : { \
-        'class' : Converter, \
-        'types' : ['hp', 'pv', 'sol_col', 'el_res', 'gas_burn'] \
+    "converter": {
+        "class": Converter,
+        "types": ["hp", "pv", "sol_col", "el_res", "gas_burn"],
     },
-    'storage' : { \
-        'class' : Storage, \
-        'types' : ['the_sto'] \
-    },
-    'distribution' : { \
-        'class' : Distribution, \
-        'types' : ['inv', 'dist_pump', 'sol_pump', 'piping'] \
+    "storage": {"class": Storage, "types": ["the_sto"]},
+    "distribution": {
+        "class": Distribution,
+        "types": ["inv", "dist_pump", "sol_pump", "piping"],
     },
 }
 
 # Edit this dictionary to add new systems
 SYSTEMS = {
-    'solar_electric' : { \
-        'class'     : System, \
-        'components' : ['hp', 'pv', 'the_sto', 'inv', 'piping', 'dist_pump'], \
-        'backup'     : ['el_res'], \
+    "solar_electric": {
+        "class": System,
+        "components": ["hp", "pv", "the_sto", "inv", "piping", "dist_pump"],
+        "backup": ["el_res"],
     },
-    'solar_thermal_new' : { \
-        'class'      : System, \
-        'components' : ['sol_col', 'the_sto', 'sol_pump', 'piping', 'dist_pump'], \
-        'backup'     : ['gas_burn'], \
+    "solar_thermal_new": {
+        "class": System,
+        "components": [
+            "sol_col",
+            "the_sto",
+            "sol_pump",
+            "piping",
+            "dist_pump",
+        ],
+        "backup": ["gas_burn"],
     },
 }
 
 # These keys are not added to the web plot
 FILTER = [
-    'hour',
-    'day',
-    'month',
-    'season',
+    "hour",
+    "day",
+    "month",
+    "season",
 ]
 
 """
@@ -74,15 +79,24 @@ Mapping class
 =================================
 """
 
+
 class Map:
-    def __init__(self, components=COMPONENTS, systems=SYSTEMS, consumer_labels=CONSUMER_LABELS, system_labels=SYSTEM_LABELS):
+    def __init__(
+        self,
+        components=COMPONENTS,
+        systems=SYSTEMS,
+        consumer_labels=CONSUMER_LABELS,
+        system_labels=SYSTEM_LABELS,
+    ):
         self.components = components
         self.systems = systems
         self._c = consumer_labels
         self._s = system_labels
 
     def __str__(self):
-        return "Components:\n{}\nSystems:\n{}".format(str(self.components), str(self.systems))
+        return "Components:\n{}\nSystems:\n{}".format(
+            str(self.components), str(self.systems)
+        )
 
     # @property
     # def components(self):
@@ -98,8 +112,8 @@ class Map:
         li = []
         for key, val in self.components.items():
             pairs = []
-            for t in val['types']:
-                pairs.append([t,self._s[t] if t in self._s else 'Undefined'])
+            for t in val["types"]:
[…]
