+++ mswh/tools/unit_converters.py
@@ -1,6 +1,7 @@
 import numpy as np
 import pandas as pd
 
+import logging
 log = logging.getLogger(__name__)
 
 
