+++ README.md
@@ -76,7 +76,7 @@ will show users how to utilize an alternative package management system in `Pyth
 
     For deactivating use `conda deactivate` or `deactivate`.
 
-6.  To make use of example `Jupyter notebooks` one should have ()`JupyterLab`)[https://jupyter.org/install] installed. To ensure the same Python kernel can be used in a `Jupyter notebook`, activate the virtual environment and run:
+6.  To make use of example `Jupyter notebooks` one should have (`JupyterLab`)[https://jupyter.org/install] installed. To ensure the same Python kernel can be used in a `Jupyter notebook`, activate the virtual environment and run:
 
         python -m ipykernel install --user --name mswh
 
