### 2020-10-26 3e90bb3
typo & added command to help mac users
