### 2020-10-31 f67b8d6
added a missing directory change that need to happen prior to pip install
@brynpickering
