### 2020-11-01 f0fd47f
switched to venv as it does allow the user to specify the python version and moved conda installation up before it shows up the first time
