+++ README.md
@@ -50,7 +50,7 @@ for further help as needed.
 ### Detailed Installation Steps
 
 This section is intended for technical users that are relatively new to virtual environments or `Python` in general, or for users who had any issues with the simple installation instructions from the previous section. This section
-will show users how to utilize an alternative package management system in `Python`, [`virtualenv`](https://packaging.python.org/guides/installing-using-pip-and-virtualenv/).
+will show users how to utilize an alternative package management system in `Python`, [`venv`](https://docs.python.org/3.8/library/venv.html).
 
 1. It is recommended to create a new `Python` environment in order to avoid interference with the system-wide Python installation, for example by using [`conda`](https://docs.conda.io/en/latest/) or [`venv`](https://docs.python.org/3.8/library/venv.html). Depending on the approach you take, pick one of the commands below and run it in a terminal to create a new environment named, for instance, `mswh`.
 
@@ -71,7 +71,7 @@ will show users how to utilize an alternative package management system in `Pyth
 
         conda activate mswh
 
-    When using `virtualenv`:
+    When using `venv`:
 
         source <path_to_env>/mswh/bin/activate
 
