### 2020-11-01 b0db1cc
Moved the standalone git lfs instruction into the detailed instructions
