### 2020-11-01 LBNL-ETA/MSWH#26 @nmstreethran
JOSS Review: Example notebook files, climate zones, and paper
[[JOSS Review](https://github.com/openjournals/joss-reviews/issues/2695)] Hi @milicag, sorry for the delay. I have a few comments regarding the example notebook files, climate zones, and paper. I will go through your submission again in the coming week and post a comment if there is anything else to be addressed. Let me know if any of my comments are unclear. Thanks.

## Example notebooks / climate zones

The `scripts` directory has been referenced in the paper and README. The "playground" notebook has not been specifically mentioned anywhere, but it demonstrates that MSWH can be extended for climates outside California. It would be nice to specify this in the README. I also think it would be good if you include a short description below the title (or somewhere else in the repository, such as the docstrings where you define `climate_zone`) and state that Banja Luka, Bosnia & Herzegovina has been included in the database as an example outside CA/US. In the contents cell, it only states that users should choose "CA climate zones available in the database".

There are a couple of typos in the notebooks, in the 'Contents' cell:

- (point 1) requirement
- (point 4) parameters

One minor comment that I have is that the acronym 'CA' is used for California in these notebooks, but this has not been defined in the first instance. Specifying what CA is when it first appears would be useful to users who are less familiar with US states.

## paper.md

- (line 51) This sentence seems to be missing a comma after 'heat pump water heater'.
- (line 62) "...model developed for in NREL's SAM software": doesn't sound quite right. I think the "in" should be removed.
- (line 64) Is "instantaneous electric gas water heater" correct? I think this should be electric resistance water heater.
- (line 81) "`@Coughlin:2020` project report": "the project report by `@Coughlin:2020`" sounds better
- (line 101) The heading has a typo (Acknowledgements)
