### 2020-11-01 LBNL-ETA/MSWH#26 @milicag
@nmstreethran Thanks, perfect timing! Great points. I'll address this in a PR together with [issue #25](https://github.com/LBNL-ETA/MSWH/issues/25).
