### 2020-11-01 6fe50b2
drafted move of detailed instructions
