### 2020-11-01 6f0e07a
moved detailed installation instructions to docs #25
