### 2020-11-01 4bb3c13
added source file for installation instructions (detailed)
