### 2020-11-01 98107b7
added link to detailed instructions
