### 2020-11-01 LBNL-ETA/MSWH#26 @milicag
Tasks:

- [x] fix language deficiencies in the paper
- [x] fix typos in the notebooks
- [x] rename playground to something straight forward,
- [x] and point out in the readme that there exist outside of CA examples in the notebooks, specifically for BL, B&H.
