### 2020-11-02 cbea764
added installation commands pertaining to the desired functionality of the example notebooks
