### 2020-11-06 LBNL-ETA/MSWH#30 @brynpickering
Backing up performance claims in JOSS paper
In the JOSS paper, you state that you `developed and implemented simplified fast performing energy balance based models`. Are you able to quantify what you mean by 'fast performing', either in a relative or absolute sense?

You could remove this claim if you are unable to quantify performance, but perhaps the better route would be to give the approximate simulation time for one of the models you have run for a past project, alongside the scale of that model (number of buildings, technologies, time scale/resolution, etc.).

Note: this issue comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/2695)
