### 2020-11-06 LBNL-ETA/MSWH#30 @brynpickering
Closing as I clearly didn't absorb the last paragraph of statement of need that covers this beautifully.
