### 2020-11-06 LBNL-ETA/MSWH#31 @brynpickering
JOSS paper structure and grammar
The content of the paper covers what I see to be the content of the software package, but it reads strangely as a publication. I would recommend restructuring and proof reading it for spelling/grammar. Some more specific comments are:

1. The summary starts with the reason for initial development ('We developed ... with a primary intent ...'). To ensure readers understand why they should use the software, could you start it with what the software is for now, rather what it was initially developed for? E.g. 'The Multiscale Solar Water Heating (MSWH) is a package to simulate individual and community scale solar water heating projects and compare them with the performance of conventional natural gas tank water heaters.'. Notice that the content is the same, but the framing is different.

2. The summary then follows as a series of sometimes disconnected statements, rather than a flow of text. The bullet points work to describe the component parts, so I wouldn't remove those, but they should be accompanied by a more coherent text. Particularly, you should avoid switching between the purpose of the software and the implementation details. I would start with a paragraph on the purpose, i.e. for comparing systems in CA and (potentially) beyond, including simple examples of things that could be compared. At this point you can briefly mention why it was initially developed to give context to e.g. the CA geographic focus. This can then be followed by the user-facing implementation, describing the existence of functional examples, a GUI (worth giving more detail on this, since it is quite a nice part of it all), and the relatively few inputs required by a user to do it all. Then you can finish on why this software is not only accessible to a user, but also functionally robust, i.e. the underlying model and its components.

3. The statement of need is more focussed on why you needed it to begin with, rather than why anybody else might need it now. As a result it is only in the last sentence where you make any mention of a target audience other than your past selves.

4. Articles tend to be missing or misused. 'MSWH' -> 'the MSWH' (and you should choose between 'the MSWH package' and 'the MSWH software'), 'SAM tool' -> 'the SAM tool', 'in order to capture sufficient level...' -> 'in order to capture a sufficient level...', 'a primary intent' -> 'the primary intent', 'very similar to model developed' -> 'very similar to the model developed', 'The application of the models' -> 'An application of the model'?

6. There are some very long sentences, which make it difficult to read. For example: 'The project that prompted the development of this software, described in @Coughlin:2020, that is at the moment of this writing in press, was enquiring whether there are any economic benefits in grouping households to be served by a single solar water heating installation, in comparison to single household solar water heating installations on the state of California level.' is a 58 word sentence.

7. Links in the PDF should point to a tagged version of the software, rather than the current master. That way, as and when the software changes, those links don't become stale. You may prefer to do this as a final step, when you have released your most up-to-date version following paper acceptance.

Note: this issue comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/2695)
