### 2020-11-06 LBNL-ETA/MSWH#32 @brynpickering
Missing contribution guidelines
The README does welcome contributions, but provides no framework doing so. [Contribution guidelines](https://docs.github.com/en/free-pro-team@latest/github/building-a-strong-community/setting-guidelines-for-repository-contributors) would be a clear way to set expectations in terms of structure and content of contributions, and help to ensure the [community can grow](https://opensource.guide/building-community/). They don't need to be as in-depth as projects like Atom, but sufficient to avoid confusion between your expectations and potential contributions.

Note: this issue comes to you as part of my wider [JOSS review](https://github.com/openjournals/joss-reviews/issues/2695)
