### 2020-11-08 LBNL-ETA/MSWH#32 @milicag
@brynpickering Let me see what I can do. What I had in mind with the section of the readme is very simple - if you'd like to contribute, you raise an issue where you describe what the issue is for, then you hear from someone on the LBL side and address the issue using a PR. Can this live on the readme, where I'd just add a few sentences, or should I pull it out into an `rst` file again? I feel like there isn't a real need for something too elaborate, as whoever raises an issue can hear back, though you might be imagining a lot more contributions than what I anticipate will happen.
