### 2020-11-08 LBNL-ETA/MSWH#31 @milicag
@brynpickering Thanks, I will address this to my best capabilities.

1 - This is as the code was a deliverable for the project. Also I though it was more honest to state that fact, but if you are so confident that this is well rounded as a standalone software then I am happy to follow your guidance.

2 - I will do my best to follow this! It's the way my brain functions.

3 - Same as 2.

4 - Typical for folks who don't have articles in their mother tongue.

5 - I developed a liking to long sentences, although I can make sure to shorten that one.

6 - Great idea.
