### 2020-11-08 LBNL-ETA/MSWH#31 @milicag
- [x] 1
- [x] 2 - Notes: A lot of the detail on components and models will be available once the report is published. They keep telling us this should happen any time now, so I would stick to referencing the report for such detail. I added some examples on future use to the statement of need section and restructured the text a bit. There were a few sentences that indeed did not make much sense any more.
- [x] 3
- [x] 4 - Notes: based [on this resource](https://english.stackexchange.com/questions/30596/using-the-definite-article-with-acronyms-and-initialisms) I will keep `SAM` rather than `the SAM`. Also `a primary intent` as this is the first time I introduce `primary intent`, otherwise you could ask `which primary intent`. For similar reasons `very similar to the model developed` should also be an `a` I think, as we are otherwise saying that something is defined which we define in the rest of the sentence, though I did change it to your liking as I can convince myself to believe in short timescale perturbations.
- [x] 5
- [x] 6 - Add to issue #16.
