### 2020-11-08 9c0e99c
addressed reviewer comments, part I
