+++ paper/paper.md
@@ -43,9 +43,9 @@ bibliography: paper.bib
 
 # Summary
 
-We developed the Multiscale Solar Water Heating (MSWH) software with a primary intent to enable users to model energy use for individual and community scale solar water heating projects and compare it with the performance of conventional natural gas tank water heaters. The package contains a [Jupyter notebook with examples](https://github.com/LBNL-ETA/MSWH/blob/master/scripts/MSWH%20System%20Tool.ipynb), a GUI (graphical user interface) developed using Django Framework and both functional and unit tests. In addition, the package is structured so that it can be extended with further technologies, applications and locations as needed.
+The Multiscale Solar Water Heating (MSWH) is a software package to simulate individual and community scale solar water heating projects and allow for a comparison with the simulation performance of conventional natural gas tank water heaters. The package contains a [Jupyter notebook with examples](https://github.com/LBNL-ETA/MSWH/blob/master/scripts/MSWH%20System%20Tool.ipynb), a GUI (graphical user interface) developed using Django Framework and both functional and unit tests. In addition, the package is structured so that it can be extended with further technologies, applications and locations as needed.
 
-We developed and implemented simplified fast performing energy balance based models. We connected the models into two preconfigured solar water heating systems that are provided with MSWH software:
+We developed and implemented simplified fast performing energy balance based models. We connected the models into two preconfigured solar water heating systems that are provided with the MSWH software:
 
 * Solar thermal collector, hot water thermal storage tank, with a selection of backups: gas storage water heater or an instantaneous gas water heater.
 * Photovoltaic panel, heat pump tank water heater, with an electric resistance water heater as backup.
@@ -59,7 +59,7 @@ The component models were either developed from scratch or implemented in Python
 * We converted natural gas tank water heater model from @Lutz:1998 into a hourly time-step model implementation.
 * Photovoltaic model is based on a simplified model found in @Wetter:2014.
 * Heat pump water heater tank is based on @Sparn:2014.
-* Solar thermal tank is a phenomenological model based on ideas very similar to model developed for NREL's SAM software [@Blair:2014].
+* Solar thermal tank is a phenomenological model based on ideas very similar to the model developed for NREL's SAM software [@Blair:2014].
 * Simplified performance data based gas burner model was implemented to represent instantaneous gas water heater.
 * Simple electric resistance model was implemented to represent instantaneous electric water heater.
 * We developed a simplified data based solar and distribution pump model.
@@ -76,13 +76,15 @@ We developed component sizing rules and size scaling rules to account for househ
 
 Performance visualization is available both in example notebooks and through the Django based GUI.
 
-The application of the models can be to explore the benefits of grouping multiple households to be served by a single solar water heating system in comparison to a system installed in a single household. Another application of the model is to calculate gas savings when changing from a gas water heater to a solar water heater in a single household.
+An application of the models can be to explore the benefits of grouping multiple households to be served by a single solar water heating system in comparison to a system installed in a single household. Another application of the model is to calculate gas savings when changing from a gas water heater to a solar water heater in a single household.
 
-More details on the hot water demand model used in creating the database of sample hot water use load profiles are freely available with the software. The software's solar radiation, component and system models for conventional natural gas and solar thermal water heating can be found in the project report by @Coughlin:2020 for which this software was developed. The software found its application in performing the engineering analysis to estimate energy consumption and savings in the project report, as well as in @Grahovac:2020. @Gerhart:2019 explains in his master's thesis the development of the GUI and provides details on the solar electric system model development. The code is also available as DOE CODE, see @Doecode:2019.
[…]
