### 2020-11-09 0f629a5
restructured, added link for future gui deployment
