+++ paper/paper.md
@@ -43,9 +43,23 @@ bibliography: paper.bib
 
 # Summary
 
-The Multiscale Solar Water Heating (MSWH) is a software package to simulate individual and community scale solar water heating projects and allow for a comparison with the simulation performance of conventional natural gas tank water heaters. The package contains a [Jupyter notebook with examples](https://github.com/LBNL-ETA/MSWH/blob/master/scripts/MSWH%20System%20Tool.ipynb), a GUI (graphical user interface) developed using Django Framework and both functional and unit tests. In addition, the package is structured so that it can be extended with further technologies, applications and locations as needed.
 
-We developed and implemented simplified fast performing energy balance based models. We connected the models into two preconfigured solar water heating systems that are provided with the MSWH software:
+The Multiscale Solar Water Heating (MSWH) is a software package to simulate individual and community scale solar water heating projects and allow for a comparison with the simulation performance of conventional natural gas tank water heaters. The package contains a [Jupyter notebook with examples](https://github.com/LBNL-ETA/MSWH/blob/master/scripts/MSWH%20System%20Tool.ipynb), a [GUI (graphical user interface) developed using Django Framework](https://solar.floweragenda.org/) and both functional and unit tests. In addition, the package is structured so that it can be extended with further technologies, applications and locations as needed. 
+
+The package was initially developed in the scope of a California Energy Commission (CEC) funded project looking at costs and benefits of using community versus individual scale solar thermal water heating systems. For this reason the database included in the MSWH software focuses primarily on California specific hot water use profiles and climate data, but can structurally accommodate any further climate zones. Scale here refers to the number of households served by a single system. Therefore, an application of the models can be to explore the benefits of grouping multiple households to be served by a single solar water heating system in comparison to a system installed in a single household. Another application of the model is to enable calculation of gas savings when switching from a gas water heater to a solar water heater in a single household.
+
+To evaluate a potential solar water heating project at design phase by looking at its simulation performance the user should create an instance of each system. For this, as is done in the detailed examples we provided in [example notebooks](https://github.com/LBNL-ETA/MSWH/tree/master/scripts), the user needs to specify the following:
+
+* The project location by choosing one of the climate zones for which data is available in our database. The database includes 16 California climate zone files and can be extended for other regions.
+* For each household: count of people supplied by the system and whether there is any daytime occupancy.
+
+We developed component sizing rules and size scaling rules to account for household occupancy and project scale, respectively. The rules are readily available in the example notebooks and can easily be modified for exploratory purposes that we further describe in the Statement of Need section. As data sources for the sizing and scaling rules we used expert knowledge, web-scraped data with the help of a tool described in @Gerke:2017 and freely available certification databases such as @Ccms:2018 and @Cec_appliance:2019 as well as sizing rules available in @Csi_thermal:2016.
+
+The MSWH software found its application in performing the engineering analysis to estimate energy consumption and savings in the @Coughlin:2020 project report, as well as in the @Grahovac:2020 research paper. @Gerhart:2019 explains in his master's thesis the development of the GUI that effectively is a flexible web framework custom built to facilitate easy addition of new system models if those are added to the MSWH software. 
+
+System performance visualization in form of time-series is available both in example notebooks and through the [Django based GUI](https://solar.floweragenda.org/).
+
+Now we would like to address some of the characteristics of the underlying models and simulation that made possible the above briefly described usage of the MSWH software. We developed and implemented simplified fast performing energy balance based component models. We connected the component models into two preconfigured solar water heating systems, that are both provided with the MSWH software. Those models are:
 
 * Solar thermal collector, hot water thermal storage tank, with a selection of backups: gas storage water heater or an instantaneous gas water heater.
 * Photovoltaic panel, heat pump tank water heater, with an electric resistance water heater as backup.
[…]
