+++ paper/paper.bib
@@ -79,12 +79,22 @@
 	pages = {19},
 }
 
+@article{DiOrio:2014,
+	title = {Technical Manual for the SAM Solar Water Heating Model},
+	language = {en},
+	journal = {National Renewable Energy Lab. (NREL), Golden, CO},
+	author = {DiOrio, Nicholas and Christensen, Craig and Burch, Jay and Dobos, Aron},
+	year = {2014},
+	url = {https://sam.nrel.gov/solar-water-heating.html},
+	pages = {24},
+}
+
 @misc{Ccms:2018,
 	title = {Compliance {Certification} {Management} {System}},
 	url = {https://www.regulations.doe.gov/ccms},
 	urldate = {2018-12-12},
 	author = {{U.S. Department of Energy - Appliance \& Equipment Standards Program}},
-  year = {2018},
+    year = {2018},
 }
 
 @misc{Cec_appliance:2019,
