+++ paper/paper.bib
@@ -3,14 +3,14 @@
 	language = {en},
 	author = {Dufﬁe, John A and Beckman, William A},
 	pages = {928},
-  publisher = {John Wiley & Sons, Inc., Hoboken, New Jersey},
-  isbn = {978-0-470-87366-3},
-  year = {2013},
+    publisher = {John Wiley & Sons, Inc., Hoboken, New Jersey},
+    isbn = {978-0-470-87366-3},
+    year = {2013},
 }
 
 @misc{Srcc:2013,
 	title = {{SRCC} {Standard} 100 - {Minimum} {Standards} {For} {Solar} {Thermal} {Collectors}},
-  author = {SRCC},
+    author = {SRCC},
 	url = {https://solar-rating.org/programs/og-100-program/},
 	urldate = {2018-10-15},
 	publisher = {Solar Rating \& Certification Corporation},
@@ -137,15 +137,15 @@
 	language = {en},
 	author = {Gerhart, Hannes},
 	pages = {82},
-  year = {2019},
-  publisher = {Technical University of Munich},
+    year = {2019},
+    publisher = {Technical University of Munich},
 }
 
 @techreport{Coughlin:2020,
 	title = {Costs and {Benefits} of {Community} versus {Individual} {End}-use {Infrastructure} for {Solar} {Water} {Heating}},
 	institution = {California Energy Commission},
 	author = {Coughlin, Katie and Grahovac, Milica and Ganeshalingam, Mohan and Hosbach, Robert and Vossos, Vagelis},
-	year = {2020, In Press}
+	year = {2020}
 }
 
 @article{Grahovac:2020,
@@ -153,6 +153,6 @@
 	language = {en},
 	author = {Grahovac, Milica and Coughlin, Katie and Ganeshalingam, Mohan and Hosbach, Robert and Vossos, Vagelis},
 	pages = {16},
-  publisher = {In Proceedings of ACEEE Summer Study 2020},
-  year = {2020},
+    publisher = {In Proceedings of ACEEE Summer Study 2020},
+    year = {2020},
 	}
