### 2020-11-11 171e00f
set footer text in settings file fixes #22
