### 2020-11-11 LBNL-ETA/MSWH#32 @brynpickering
I suppose I hope for many more contributions than you anticipate! A short contribution guideline is also fine, but the flow you give in your comment here is not really given in the README. What I see here is that you expect interested parties to *first* submit an issue, then the process of addressing it with a PR can be discussed in the issue. You could describe this and also give a brief description of the kind of issues you welcome (bug reports, feature requests, etc.), for those who don't really know what's going on.

You could also think about adding [issue and PR templates](https://docs.github.com/en/free-pro-team@latest/github/building-a-strong-community/configuring-issue-templates-for-your-repository), if you want to ensure all issues and PRs follow some consistent structure.
