### 2020-11-11 LBNL-ETA/MSWH#32 @milicag
Thank you for the thought. I will add a description to the readme similar to what I wrote in my response. I will consider a template should there start being significant contributions in the future.
