### 2020-11-12 #2695 @brynpickering
@sjpfenninger, we ended up moving discussion to the software repo, but I can give an update from my side. Overall, there's a lot of content in the submission, but I have become unsure as to whether it fits within the scope of JOSS. This concern aside, I have no new issues to raise, and see my review of this submission to be close to completion. Below are more specific comments which are being / have been addressed in individual issues.

Software
---------
The code looks to be sound, and the tests are thorough, if not entirely comprehensive; triggering of several error/info messages is not being caught by any tests, but core mathematical functionality is well tested.  The implementation could be considered a bit impenetrable for would-be contributors, since variable names aren't that human-readable.

Documentation
---------------
There is API documentation, and a Jupyter notebook as an example of a model, but otherwise the documentation is quite sparse. It probably meets the minimum JOSS requirements, but makes no attempt to go into detail on what's going on in the internals, nor to more thoroughly document the vast array of input parameters expected from the user (most of which have default values). We've gone in circles about installation instructions, but seem to have reached a point now where everyone is happy.

Paper
------
Following my review (https://github.com/LBNL-ETA/MSWH/issues/31), the paper has recently undergone a major revision, which I will ask Whedon to re-generate following the merging of the changes. From a quick look, I expect to only make minor comments on this update to the manuscript.

Current concern
----------------
The software is open-source and well written, but it seems to be an archive of a piece of software produced for a specific (past) project. The original focus of the paper (https://github.com/LBNL-ETA/MSWH/issues/31), and a separate conversation about contribution guidelines (https://github.com/LBNL-ETA/MSWH/issues/32), raises the question of whether there is any expectation that a community grows around developing this software further. The past project(s) focussed on California, and although it is technically feasible to model anywhere in the world, the existing parametrisation and data processing retains that geographic focus. From the PR for an update to the paper:

> The water consumption profiles can be highly location specific and their development for additional climate zones would require new research efforts, although a quick approximation may be made with caution by scaling the California profiles to match the location specific estimate of the average annual water use.

**I would appreciate it if this concern could be addressed by the editorial team**. Perhaps it shouldn't be a concern at all.
