### 2020-11-12 9a5c45d
elaborated on additional climate inputs
