### 2020-11-12 #2695 @milicag
@sjpfenninger @sjpfenninger @nmstreethran I'm happy to address any outstanding issues!

Regarding the current concern section, I added a sentence to [the current PR branch](https://github.com/LBNL-ETA/MSWH/pull/33) that makes the quoted out section of the paper sound more as I had intended - the section is referring to future research. I don't think that any developer of software can provide all the possible inputs and for that reason I do give a bit of a guidance in the paper. It is essentially a tip to the user who does not reside in California. For that reason there is also an example for Bosnia. The paper by JOSS approach should be really short. The journal is, what draw me towards it, dedicated to software itself. I believe we should refocus our efforts here on the software itself.

All variables that are user facing are described in the code documentation, in each method or class. The code would be very long and hard to understand from the coding point of view if I was to use full words for each variable. I hope you understand this.

The software package is complete and that is also why I decided it is ready to be published as such. Any future research depends on the available funding and I cannot see how this relates to current state and JOSS publication. I would certainly love to have future contributions!

Please feel free to provide comment and I am here to clarify anything you might still be wondering about.

Thanks,
Milica
