### 2020-11-13 LBNL-ETA/MSWH#32 @milicag
@brynpickering What would be your preference, also from the JOSS perspective, on adding the contribution guidelines?
