### 2020-11-13 LBNL-ETA/MSWH#32 @sjpfenninger
@milicag I think these can be relatively minimal, following examples given in the [GitHub documentation](https://docs.github.com/en/free-pro-team@latest/github/building-a-strong-community/setting-guidelines-for-repository-contributors)
