### 2020-11-13 #2695 @sjpfenninger
@milicag @brynpickering Thanks both for the update. I see no major issue here. @milicag You could explicitly point to the California-specific data as examples included, but they do not preclude application of the software elsewhere. @nmstreethran Could you provide an update what is outstanding from your side?
