### 2020-11-13 LBNL-ETA/MSWH#31 @brynpickering
The changes look good to me. Just some minor comments:

1. I certainly hope it is valuable as a stand-alone software. A lot of time and effort has gone into it, so it would be great if others use and extend this software.

2. The structure looks good to me; it read much more fluidly. The only bit I would look to update is the sentences starting `Now we would like to address some of the characteristics`, which still read a bit strangely.

4. On the `SAM` question, since you're referring to the `tool` and `SAM` is something of a [noun adjunct](https://en.wikipedia.org/wiki/Noun_adjunct) (grammar buffs might correct me), it should be `the SAM tool`. If referring to just `SAM` then the `the` is indeed redundant. Happy to concede on `primary`. I guess I see it as a superlative, so there can only be one primary thing (and therefore = `the primary ...`), but that's probably a questionable viewpoint ;). Also happy with either article for `model`, it just had neither before.

5. Long sentences are a  dangerous thing to become addicted to, best to wean yourself off them sooner rather than later!

I have seen a number of minor grammatical errors creep back into the text on this re-write. Are you able to have a native English speaker read through it and make necessary edits?
