### 2020-11-13 LBNL-ETA/MSWH#31 @milicag
@rhosbach Would you have time to take a quick look at this version of the paper and see if there is anything that looks grammatically incorrect from your point of view?
