### 2020-11-13 LBNL-ETA/MSWH#31 @milicag
@brynpickering Are there any other action items on this or may I merge after Robert takes a look? Would you like to read it again after Robert finishes or should that be considered finalized at that point?

Many thanks!
