### 2020-11-13 LBNL-ETA/MSWH#31 @milicag
@brynpickering: I am hoping that Robert will be willing and have time to take a look. The team of co-authors has reviewed the paper prior to the submission. I live for many years now in an English speaking country. I do not understand what is wrong with starting a sentence with: `Now we would like to address some of the characteristics`. The international scientific community should respect the conventions put forth by diversity and inclusion.
