### 2020-11-14 d0653a2
added contributing guidelines for #32, JOSS review
