+++ contributing.md
@@ -1,0 +1,26 @@
+# Guidelines for contributors
+
+The MSWH software is open to contributions from the coder community around the world. 
+
+The contributors should comply with the 
+[University of California Standards of Ethical Conduct](https://policy.ucop.edu/doc/1100172/EthicalValuesandConduct).
+
+The contribution can be a bug fix, an improvement to existing features, or a fully new feature.
+
+If you wish to contribute any of these please follow this workflow:
+
+* Create an new issue in the [MSWH issue tracker](https://github.com/LBNL-ETA/MSWH/issues) that briefly describes your intended contribution
+* Await response from a person responsible for code maintenance
+* Discuss whether there is a need to address this issue and ways in which the issue should best be addressed
+* Once agreement is reached on how to address the issue, develop the feature, fix the bug or improve the code
+* When you are happy with your code that addresses the issue, after you confirm that all tests are passing, create a [pull request](https://github.com/LBNL-ETA/MSWH/compare) against the `master` branch
+* Notify the person that was discussing the issue with you using `@githubname` that the feature/fix/improvement is now ready for review
+* Conduct iterations of receiving review and addressing it until the reviewer approves the PR.
+* The reviewer will then thank you and merge the PR
+* Now you can start utilizing the updated code!
+
+In addition, if you have question on how to best apply the existing code for your particular purpose, you may pose that question using the [MSWH issue tracker](https://github.com/LBNL-ETA/MSWH/issues).
+
+Thank you for your interest in contributing to our open source MSWH repository!
+
+Milica
