### 2020-11-15 #2695 @nmstreethran
Hi @sjpfenninger, sorry for the late response.

Based on my checklist, I only have one item remaining, which is a statement of need in the documentation. I think just a summary of the target audience would suffice as the usage of the software has already been described in the documentation [here](https://lbnl-eta.github.io/MSWH/source/models.html). I have checked off all other items as I believe PRs https://github.com/LBNL-ETA/MSWH/pull/33 and https://github.com/LBNL-ETA/MSWH/pull/35 will resolve them.

I opened issues https://github.com/LBNL-ETA/MSWH/issues/13, https://github.com/LBNL-ETA/MSWH/issues/14, https://github.com/LBNL-ETA/MSWH/issues/15, and https://github.com/LBNL-ETA/MSWH/issues/26 as part of my review, which have all been resolved. I can confirm that the software installation, example notebooks, web application, and tests run as expected.

Overall, I think this is a useful piece of software and I am happy to recommend to accept this submission once the outstanding PRs have been merged and the documentation is complete.
