+++ contributing.md
@@ -7,17 +7,16 @@ The contributors should comply with the
 
 The contribution can be a bug fix, an improvement to existing features, or a fully new feature. If you wish to contribute any of those please follow this workflow:
 
+1. Fork the MSWH repository.
 1. Create an new issue in the [MSWH issue tracker](https://github.com/LBNL-ETA/MSWH/issues) that briefly describes your intended contribution.
 1. Await response from a person responsible for code maintenance.
-1. Discuss whether there is a need to address this issue and ways in which the issue should best be addressed.
-1. Once agreement is reached on how to address the issue, develop the code on a new branch that includes the issue number for easy tracking.
-1. When you are happy with your code that addresses the issue, after you confirm that all tests are passing, create a [pull request](https://github.com/LBNL-ETA/MSWH/compare) against the `master` branch.
+1. Discuss whether there is a need to address this issue and ways in which the issue should be best  addressed.
+1. Once agreement is reached on how to address the issue, develop the code on a new branch on your forked repository. It is highly recommended that the branch includes the MSWH software repo issue number for easier tracking.
+1. When you are happy with your code that addresses the issue, after you confirm that all tests are passing, create a [pull request](https://github.com/LBNL-ETA/MSWH/compare) against the `master` branch of the MSWH repository.
 1. Notify the person that was discussing the issue with you using `@githubname` that the feature/fix/improvement is now ready for review.
 1. Conduct iterations of receiving review and addressing it until the reviewer approves the PR.
 1. The reviewer will then thank you and merge the PR.
 1. Now you can start utilizing the updated code!
-
-Alternatively you may also fork our repository and develop the new feature there. If you believe that we might be interested in your new feature you can follow the same steps, with a difference in step 4. Namely, you will create a new branch on the fork of our repository and propose a PR using that branch.
 
 In addition, if you have questions on how to best apply the existing code for your particular purpose, you may pose that question using the [MSWH issue tracker](https://github.com/LBNL-ETA/MSWH/issues).
 
