### 2020-11-15 f15443f
Fixed typos and errors
