### 2020-11-15 bf11bd1
Updated contribution guidelines link
