+++ paper/paper.md
@@ -57,7 +57,7 @@ We developed component sizing rules and size scaling rules to account for the ho
 
 The MSWH software found its application in performing the engineering analysis to estimate energy consumption and savings in the @Coughlin:2020 project report, as well as in the @Grahovac:2020 research paper. @Gerhart:2019 is a master's thesis that explains the development of [the GUI](https://github.com/LBNL-ETA/MSWH/tree/master/web) and how to effectively use the GUI as a flexible web framework, since it was custom built to facilitate easy addition of new system models when those are added to the MSWH software analogously to already existing ones. 
 
-System performance visualization in form of time-series is available both in example notebooks and through the GUI, either spined of locally or [using a web deployed version](https://solar.floweragenda.org/).
+System performance visualization in form of time-series is available both in example notebooks and through the GUI, either spined off locally or [using a web deployed version](https://solar.floweragenda.org/).
 
 Now we would like to address some of the characteristics of the underlying models and simulation that made possible the above briefly described usage of the MSWH software. We developed and implemented simplified fast performing energy balance based component models. We connected the component models into two preconfigured solar water heating systems, that are both provided with the MSWH software. Those models are:
 
@@ -110,7 +110,7 @@ When it comes to future application of the MSWH software, we can envision four m
 * Homeowners.
 * Educators.
 
-If the features the existing MSWH software are sufficient for their application, the policy developers and researchers could utilize the existing MSWH software by embedding it into some larger analysis framework they construct such that it provides answers to their specific research questions. Should they have a need for additional system configurations and even additional components, the existing framework should be expanded in line with the structure made available to the user in the MSWH software. When systems are added following the structure of the existing systems, an addition of such new system to the GUI is made possible by using the flexible web framework. 
+If the features of the existing MSWH software are sufficient for their application, the policy developers and researchers could utilize the existing MSWH software by embedding it into some larger analysis framework they construct such that it provides answers to their specific research questions. Should they have a need for additional system configurations and even additional components, the existing framework should be expanded in line with the structure made available to the user in the MSWH software. When systems are added following the structure of the existing systems, an addition of such new system to the GUI is made possible by using the flexible web framework. 
 
 For those who in their professional life deal with planning, designing and contracting of the solar thermal water heating systems it might be useful to have an access to a freely available simulation tool, such as the MSWH software, that they can use to evaluate various system designs. The design parameters that such users can easily modify are household occupancies, climate zone, collector and tank sizes, component performance parameters such as insulation level of any thermal storage tanks, and types of solar collectors. The MSWH software relies on standard collector rating data readily available for most designs found on the market today. For each proposed design the MSWH software will output, among other results, the solar fraction and the backup energy use on an annual level, the two variables allowing for a quick cross-comparison for the proposed designs.
 
