### 2020-11-15 #2695 @milicag
@sjpfenninger At @brynpickering's suggestion we are having a native speaker taking a final look at the `paper.md` on Monday and with that the current pull request will provide all edits the reviewers identified as needed.

I have put together [this finalization issue](https://github.com/LBNL-ETA/MSWH/issues/16) that I intend to address as the last step, given your approval.
