### 2020-11-16 LBNL-ETA/MSWH#31 @brynpickering
@milicag my comment on having a native English speaker go through it was due to your own admission that certain mistakes could be attributed to English not being your mother tongue. Except for a few slip-ups, your English is very good. It would just be good to iron out those few issues before publication.

The paper structure and writing style is independent of the grammar or language profiency. In this case, having sentences that are to the point will improve clarity. Although not particularly academic, there's nothing wrong with a sentence starting `Now we would like to address some of the characteristics...`, it's just redundant. It would be enough to simply state "The MSWH software connects simplified, fast-performing component models to describe two preconfigured solar water heating systems. These systems and their component models are: ...".

"quality of writing" is one of the JOSS paper checkpoints, and I had hoped to point to an area that would benefit from revisiting. Everyone wins when papers are as readable as possible! Saying that, I won't push this further, if you and your co-authors would prefer not to spend more time on it.
