### 2020-11-16 LBNL-ETA/MSWH#31 @milicag
@brynpickering I am certainly happy to work towards meeting the required quality of writing!
