### 2020-11-16 LBNL-ETA/MSWH#31 @brynpickering
> @brynpickering If you allow me to have a say in this, to me saying "The MSWH software connects" lacks accuracy and is from an academic perspective misleading, as we have connected the components in systems manually. The tools that do such connections automatically are hard to build. Would it be fine with you if I rephrase it in some other way?

Of course! My suggestions are just that. You are best placed to rephrase in order to retain the desired meaning.
