### 2020-11-16 LBNL-ETA/MSWH#31 @milicag
I just offered a modification, I think it reads more academic. Please let me know what you think. If there are any other sentences that you wish were more clear, please point them out. I appreciate your level of involvement and am very happy to polish up the paper this way.
