### 2020-11-16 LBNL-ETA/MSWH#31 @milicag
From @rhosbach: Mostly the review contains alternate wordings for sentences. Feel free to edit/change as you see fit.

@brynpickering, @nmstreethran: I will address Robert's review later tonight. Please let me know if you need anything else.
