### 2020-11-17 13e55a3
updated first latter formatting in the references
