+++ paper/paper.md
@@ -53,9 +53,9 @@ To evaluate a solar water heating project at the design phase by looking at its 
 * The project location by choosing one of the climate zones for which data is available in our database. The database includes 16 California climate zone files and can be extended to other regions.
 * For each household: count of people supplied by the system and whether there is any daytime household occupancy.
 
-We developed component sizing rules and size scaling rules to account for the household occupancy and project scale, respectively. The rules are readily available in the example notebooks and can easily be modified for exploratory purposes that we further describe in the Statement of Need section. For the sizing and scaling rules, we used the following data sources: expert knowledge, web-scraped data with the help of a tool described in @gerke:2017, sizing rules available in @Csi_thermal:2016, and certification databases such as @ccms:2018 and @Cec_appliance:2019.
+We developed component sizing rules and size scaling rules to account for the household occupancy and project scale, respectively. The rules are readily available in the example notebooks and can easily be modified for exploratory purposes that we further describe in the Statement of Need section. For the sizing and scaling rules, we used the following data sources: expert knowledge, web-scraped data with the help of a tool described in @Gerke:2017, sizing rules available in @Csi_thermal:2016, and certification databases such as @Ccms:2018 and @Cec_appliance:2019.
 
-The MSWH software was used to perform the engineering analysis to estimate energy consumption and savings in the @coughlin:2020 project report as well as in the @grahovac:2020 research paper. @gerhart:2019 is a master's thesis that explains the development of [the GUI](https://github.com/LBNL-ETA/MSWH/tree/master/web) and how to effectively use the GUI as a flexible web framework, because it was custom-built to facilitate easy addition of new system models when those are added to the MSWH software.
+The MSWH software was used to perform the engineering analysis to estimate energy consumption and savings in the @Coughlin:2020 project report as well as in the @Grahovac:2020 research paper. @Gerhart:2019 is a master's thesis that explains the development of [the GUI](https://github.com/LBNL-ETA/MSWH/tree/master/web) and how to effectively use the GUI as a flexible web framework, because it was custom-built to facilitate easy addition of new system models when those are added to the MSWH software.
 
 System performance time series visualizations are available both in example notebooks and through the GUI, either spun off locally or [using a web deployed version](https://solar.floweragenda.org/).
 
