### 2020-11-17 #2695 @milicag
@brynpickering and @nmstreethran I just merged into the main branch the two PRs in response to the latest reviewer comment.

[This PR](https://github.com/LBNL-ETA/MSWH/pull/35) brought in the contribution guidelines and added a short statement of need to the README file. After Robert, one of the co-authors provided a review I introduced his suggested edits and merged into main through this [PR](https://github.com/LBNL-ETA/MSWH/pull/33). I compiled the paper using your preview capabilities.

[This issue](https://github.com/LBNL-ETA/MSWH/issues/16) summarizes the remaining steps I would like to take next.

Please let me know if there is anything additional that needs to be done.
