### 2020-11-18 #2695 @milicag
> @milicag @brynpickering Thanks both for the update. I see no major issue here. @milicag You could explicitly point to the California-specific data as examples included, but they do not preclude application of the software elsewhere. @nmstreethran Could you provide an update what is outstanding from your side?

I believe this is well addressed in the revised paper, however, do let me know if you find that it needs additional explanation in the paper.
