### 2020-11-18 #2695 @milicag
@brynpickering @nmstreethran I will wait for your prompt to start addressing [the remaining issue](https://github.com/LBNL-ETA/MSWH/issues/16), as I'd like to tag a code release that goes with the JOSS paper, to avoid a need to re-tag.
