### 2020-11-18 LBNL-ETA/MSWH#16 @milicag
- Paper: spell out 1 in "1 hour".
