### 2020-11-19 #2695 @brynpickering
@milicag thanks for your work on this, I'm happy with the result and have completed my review checklist.

A minor point (which I think you brought up in a now-closed issue/PR) is that the section of your summary relating to the components of the model could be placed in a sub-section (e.g. 'model components'), to break up the summary slightly. Similarly, a sub-section for 'future applications'/'user groups' in the Statement of Need section would work well. Not sure if this violates some strict structure that JOSS papers must adhere to (@sjpfenninger?), though.
