### 2020-11-19 #2695 @milicag
@brynpickering Thank you! This is good news for me and the team. The paper is pretty short, only four pages of text, so I don't see a real reason to break it into the subsections, though I also see how that would help if someone is to read the paper really quickly. I leave it to @sjpfenninger.
