### 2020-11-19 #2695 @milicag
@brynpickering, @nmstreethran - Thank you for your thorough review, it has improved many aspects of the software, especially the accessibility to users and contributors! I appreciate all your help, suggestions, and all our discussions. This was a really unique experience in these pandemic times.
