### 2020-11-22 LBNL-ETA/MSWH#16 @milicag
Compiled and sorted list of remaining tasks:

- [x] Paper: spell out 1 in "1 hour"
- [x] Hannes third affiliation - remove it
- [x] dedicate the paper to @kraskrom because, among other aspects, he can write any equation upside down and this issue number is 16
- [x] in the `paper.md` update the hyperlinks to point to release 2.0.0 files rather than master
- [x] In `models.rst` add a sentence in the Sphinx documentation telling users that development primarily takes place on the GitHub repository, and that they can refer to the contributing guidelines and README for instructions on how to run tests
- [x] make sure the pointers on the gh-pages make sense
- [x] regenerate gh-pages and the pdf of the documentation, check all links - joss paper, docstrings, readme
- [x] merge, test code, test example notebooks
- [x] tag the codebase with release 2.0.0
- [x] test links
- [x] Follow the remaining steps on https://joss.readthedocs.io/en/latest/submitting.html#the-review-process
