### 2020-11-23 ed818bc
addressed remaining items in the paper text, added contributing section to models.rst
