+++ paper/paper.md
@@ -23,8 +23,8 @@ authors:
     affiliation: "1"
   - name: Katie Coughlin
     affiliation: "1"
-  - name: Hannes Gerhart^[at the time of code creation was at affiliation 1 and 2, now is at affiliation 3]
-    affiliation: "1, 2, 3"
+  - name: Hannes Gerhart^[At the time of code creation was at affiliation 1 and 2]
+    affiliation: "1, 2"
   - name: Robert Hosbach
     affiliation: "1"
 
@@ -33,8 +33,6 @@ affiliations:
     index: 1
   - name: Technical University of Munich, Munich, Germany
     index: 2
-  - name: Cloudflare, Inc.
-    index: 3
 
 date: September 7, 2020
 bibliography: paper.bib
@@ -44,18 +42,18 @@ bibliography: paper.bib
 # Summary
 
 
-The Multiscale Solar Water Heating (MSWH) package simulates individual and community scale solar water heating projects and allows for a comparison with the simulation performance of conventional natural gas tank water heaters. The package contains a [Jupyter notebook with examples](https://github.com/LBNL-ETA/MSWH/blob/master/scripts/MSWH%20System%20Tool.ipynb), a [graphical user interface (GUI) developed using Django Framework](https://github.com/LBNL-ETA/MSWH/tree/master/web) and both functional and unit tests. Also, the package is structured so that it can be extended with further technologies, applications, and locations as needed.
+The Multiscale Solar Water Heating (MSWH) package simulates individual and community scale solar water heating projects and allows for a comparison with the simulation performance of conventional natural gas tank water heaters. The package contains a [Jupyter notebook with examples](https://github.com/LBNL-ETA/MSWH/blob/v2.0.0/scripts/MSWH%20System%20Tool.ipynb), a [graphical user interface (GUI) developed using Django Framework](https://github.com/LBNL-ETA/MSWH/tree/v2.0.0/web) and both functional and unit tests. Also, the package is structured so that it can be extended with further technologies, applications, and locations as needed.
 
 The package was initially developed in the scope of a California Energy Commission (CEC) funded project looking at costs and benefits of using community versus individual scale solar thermal water heating systems. For this reason, the database included in the MSWH software focuses primarily on California-specific hot water use profiles and climate data, but can structurally accommodate any further climate zones. The scale here refers to the number of households served by a single system. Therefore, one can apply the models to explore the benefits of grouping multiple households to be served by a single solar water heating system in comparison to a system installed in a single household. Another example application of the models is to enable calculation of gas savings when switching from a gas water heater to a solar water heater in a single household.
 
-To evaluate a solar water heating project at the design phase by looking at its simulation performance the user should create a system instance for each compared system. For this, as is done in the detailed examples we provided in [example notebooks](https://github.com/LBNL-ETA/MSWH/tree/master/scripts), the user needs to specify the following:
+To evaluate a solar water heating project at the design phase by looking at its simulation performance the user should create a system instance for each compared system. For this, as is done in the detailed examples we provided in [example notebooks](https://github.com/LBNL-ETA/MSWH/tree/v2.0.0/scripts), the user needs to specify the following:
 
 * The project location by choosing one of the climate zones for which data is available in our database. The database includes 16 California climate zone files and can be extended to other regions.
 * For each household: count of people supplied by the system and whether there is any daytime household occupancy.
 
 We developed component sizing rules and size scaling rules to account for the household occupancy and project scale, respectively. The rules are readily available in the example notebooks and can easily be modified for exploratory purposes that we further describe in the Statement of Need section. For the sizing and scaling rules, we used the following data sources: expert knowledge, web-scraped data with the help of a tool described in @Gerke:2017, sizing rules available in @Csi_thermal:2016, and certification databases such as @Ccms:2018 and @Cec_appliance:2019.
 
[…]
