### 2020-11-23 58beeaa
edits related to joss issue #16
