### 2020-11-23 #2695 @milicag
The version number: v2.0.0
DOI: 10.5281/zenodo.4285623

@sjpfenninger
