### 2020-11-23 3900ccf
acknowledged JOSS publication contributors
