### 2020-11-30 #2695 @sjpfenninger
Thanks @nmstreethran and @brynpickering for your extensive reviews.

@milicag: We are nearing completion on this. However, your paper is on the long side. We ask for papers to be 250-1000 words long and the more extensive information should really live in the project's documentation:

I think this can be addressed with two easy changes:

1. In the Summary section, move everything between the paragraphs starting with "We developed component sizing rule" and "The MSWH software is both accessible" into the documentation, with the exception of the paragraph starting with "The MSWH software was used to perform".

2. Move the statement of need to the documentation (maybe merging with https://lbnl-eta.github.io/MSWH/source/models.html, as suggested by @brynpickering), keeping the first three paragraphs (up to the one starting with "Modelica buildings library") in the paper, and maybe adding a fourth paragraph that summarises anything important from the remainder of the removed section including future applications and possible users.
