### 2020-12-01 fbb98c8
drafted edits, moved text, added tasks
