+++ paper/paper.md
@@ -51,39 +51,13 @@ To evaluate a solar water heating project at the design phase by looking at its 
 * The project location by choosing one of the climate zones for which data is available in our database. The database includes 16 California climate zone files and can be extended to other regions.
 * For each household: count of people supplied by the system and whether there is any daytime household occupancy.
 
-We developed component sizing rules and size scaling rules to account for the household occupancy and project scale, respectively. The rules are readily available in the example notebooks and can easily be modified for exploratory purposes that we further describe in the Statement of Need section. For the sizing and scaling rules, we used the following data sources: expert knowledge, web-scraped data with the help of a tool described in @Gerke:2017, sizing rules available in @Csi_thermal:2016, and certification databases such as @Ccms:2018 and @Cec_appliance:2019.
-
 The MSWH software was used to perform the engineering analysis to estimate energy consumption and savings in the @Coughlin:2020 project report as well as in the @Grahovac:2020 research paper. @Gerhart:2019 is a master's thesis that explains the development of [the GUI](https://github.com/LBNL-ETA/MSWH/tree/v2.0.0/web) and how to effectively use the GUI as a flexible web framework, because it was custom-built to facilitate easy addition of new system models when those are added to the MSWH software.
-
-System performance time series visualizations are available both in example notebooks and through the GUI, either spun off locally or [using a web deployed version](https://solar.floweragenda.org/).
-
-In this section we briefly introduce the characteristics of the underlying models and simulation. We developed and implemented simplified fast performing energy balance based component models. We connected the component models into two preconfigured solar water heating systems, that are both provided with the MSWH software. Those models are:
-
-* Solar thermal collector, hot water thermal storage tank, with a selection of backups: gas storage water heater or an instantaneous gas water heater.
-* Photovoltaic panel, heat pump tank water heater, with an electric resistance water heater as backup.
-
-We built a simple simulation solver that uses the explicit forward Euler method to solve the balance equations in each simulation time-step.
-
-The component models were either developed from scratch or implemented in Python based on existing models identified in the literature. We implemented the following existing or new models:
-
-* Solar irradiation on a tilted surface model is based on equations found in @Duffie:2013.
-* Solar collector models and model parameters are based on @Ashrae:2013 and @Srcc:2013.
-* We converted the natural gas tank water heater model from @Lutz:1998 into an hourly time-step model implementation.
-* Photovoltaic model is based on a simplified model found in @Wetter:2014.
-* Heat pump water heater tank is based on @Sparn:2014.
-* Solar thermal tank is a phenomenological model based on ideas very similar to the model developed for NREL's SAM software (@Blair:2014), as described in @DiOrio:2014.
-* Simplified performance data-based gas burner model was implemented to represent instantaneous gas water heater.
-* Simple electric resistance model was implemented to represent instantaneous electric water heater.
-* We developed a simplified data based solar and distribution pump model.
-* To model the distribution piping network we developed a simplified model that is capable of accounting for thermal losses at stagnation and flows on-demand with correction factors available to help account for the relatively long time-step of one hour.
-
-More details on the hot water demand model used in creating the database of sample hot water use load profiles, as well as extensive detail on the software's solar radiation, component and system models can be found in the project report by @Coughlin:2020. @Gerhart:2019 thesis provides additional details on the solar electric system model development.
-
[…]
