+++ docs/source/models.rst
@@ -3,6 +3,89 @@
 Multiscale Solar Water Heating (MSWH)
 -------------------------------------
 
+
+
+==================  from paper
+
+We developed component sizing rules and size scaling rules to account for the household occupancy and project scale, respectively. The rules are readily available in the example notebooks and can easily be modified for exploratory purposes that we further describe in the Statement of Need section. For the sizing and scaling rules, we used the following data sources: expert knowledge, web-scraped data with the help of a tool described in @Gerke:2017, sizing rules available in @Csi_thermal:2016, and certification databases such as @Ccms:2018 and @Cec_appliance:2019.
+
+System performance time series visualizations are available both in example notebooks and through the GUI, either spun off locally or [using a web deployed version](https://solar.floweragenda.org/).
+
+In this section we briefly introduce the characteristics of the underlying models and simulation. We developed and implemented simplified fast performing energy balance based component models. We connected the component models into two preconfigured solar water heating systems, that are both provided with the MSWH software. Those models are:
+
+* Solar thermal collector, hot water thermal storage tank, with a selection of backups: gas storage water heater or an instantaneous gas water heater.
+* Photovoltaic panel, heat pump tank water heater, with an electric resistance water heater as backup.
+
+We built a simple simulation solver that uses the explicit forward Euler method to solve the balance equations in each simulation time-step.
+
+The component models were either developed from scratch or implemented in Python based on existing models identified in the literature. We implemented the following existing or new models:
+
+* Solar irradiation on a tilted surface model is based on equations found in @Duffie:2013.
+* Solar collector models and model parameters are based on @Ashrae:2013 and @Srcc:2013.
+* We converted the natural gas tank water heater model from @Lutz:1998 into an hourly time-step model implementation.
+* Photovoltaic model is based on a simplified model found in @Wetter:2014.
+* Heat pump water heater tank is based on @Sparn:2014.
+* Solar thermal tank is a phenomenological model based on ideas very similar to the model developed for NREL's SAM software (@Blair:2014), as described in @DiOrio:2014.
+* Simplified performance data-based gas burner model was implemented to represent instantaneous gas water heater.
+* Simple electric resistance model was implemented to represent instantaneous electric water heater.
+* We developed a simplified data based solar and distribution pump model.
+* To model the distribution piping network we developed a simplified model that is capable of accounting for thermal losses at stagnation and flows on-demand with correction factors available to help account for the relatively long time-step of one hour.
+
+More details on the hot water demand model used in creating the database of sample hot water use load profiles, as well as extensive detail on the software's solar radiation, component and system models can be found in the project report by @Coughlin:2020. @Gerhart:2019 thesis provides additional details on the solar electric system model development.
+
+Note that the weather data are currently mostly limited to California and can be extended to other climate zones. An example climate zone outside of California was added for Banja Luka, Bosnia and Herzegovina, through an [additional example Jupyter notebook](https://github.com/LBNL-ETA/MSWH/blob/v2.0.0/scripts/MSWH&#32;System&#32;Tool&#32;-&#32;Additional&#32;Climate.ipynb). The water consumption profiles can be highly location specific and their development for additional climate zones would require new research efforts. A quick approximation may be made with caution by scaling the California profiles to match the location-specific estimate of the average annual water use. This is possible as the shape of each daily profile can be assumed similar and sufficiently variable to allow for the study of transient and peak load effects at any location. The weather processor is TMY3 enabled and the user may populate the database with additional climates as needed.
+
+
+===============  From the statement of need from the paper:
+
[…]
