### 2020-12-02 d2f831d
drafted edits to documentation and paper, with bibliography
