+++ paper/paper.bib
@@ -1,116 +1,9 @@
-@article{Duffie:2013,
-	title = {Solar {Engineering} of {Thermal} {Processes}},
-	language = {en},
-	author = {Dufﬁe, John A and Beckman, William A},
-	pages = {928},
-    publisher = {John Wiley & Sons, Inc., Hoboken, New Jersey},
-    isbn = {978-0-470-87366-3},
-    year = {2013},
-}
-
-@misc{Srcc:2013,
-	title = {{SRCC} {Standard} 100 - {Minimum} {Standards} {For} {Solar} {Thermal} {Collectors}},
-    author = {SRCC},
-	url = {https://solar-rating.org/programs/og-100-program/},
-	urldate = {2018-10-15},
-	publisher = {Solar Rating \& Certification Corporation},
-	month = jan,
-	year = {2013},
-}
-
-@misc{Ashrae:2013,
-	title = {ASHRAE 93:2003 - {Method} {Of} {Testing} {To} {Determine} {The} {Thermal} {Performance} {Of} {Solar} {Collectors}},
-  author = {ASHRAE},
-	publisher = {American Society of Heating, Refrigerating \& Air Conditioning Engineer},
-	year = {2013}
-}
-
-@inproceedings{Lutz:1998,
-	title = {{WHAM}: {A} {Simplified} {Energy} {Consumption} {Equation} for {Water} {Heaters}},
-	volume = {1},
-	shorttitle = {{WHAM}: {A} {Simplified} {Energy} {Consumption} {Equation} for {Water} {Heaters}},
-	url = {http://aceee.org/files/proceedings/1998/data/papers/0114.PDF},
-	language = {English},
-	urldate = {2015-10-19},
-	booktitle = {472246},
-	publisher = {ACEEE},
-	author = {Lutz, James D. and Dunham Whitehead, Camilla and Lekov, Alex and Winiarski, David and Rosenquist, Gregory},
-	month = mar,
-	year = {1998},
-	keywords = {Appliance Standards Water Heaters},
-	pages = {171--183},
-}
-
-@article{Wetter:2014,
-	title = {Modelica {Buildings} library},
-	volume = {7},
-	issn = {1940-1493, 1940-1507},
-	url = {http://simulationresearch.lbl.gov/modelica/},
-	doi = {10.1080/19401493.2013.765506},
-	journal = {Journal of Building Performance Simulation},
-	author = {Wetter, Michael and Zuo, Wangda and Nouidui, Thierry S. and Pang, Xiufeng},
-	month = jul,
-	year = {2014},
-	pages = {253--270},
-	file = {Wetter et al. - 2014 - Modelica Buildings library.pdf:/Users/milicagrahovac/Zotero/storage/AJ9A96ZP/Wetter et al. - 2014 - Modelica Buildings library.pdf:application/pdf}
-}
-
-@article{Gerke:2017,
-	title = {The {International} {Database} of {Efficient} {Appliances} ({IDEA}): {A} new tool to support appliance energy-efficiency deployment},
-	volume = {205},
-	issn = {0306-2619},
-	shorttitle = {The {International} {Database} of {Efficient} {Appliances} ({IDEA})},
-	url = {http://www.sciencedirect.com/science/article/pii/S0306261917309650},
-	doi = {10.1016/j.apenergy.2017.07.093},
-	journal = {Applied Energy},
-	author = {Gerke, Brian F. and McNeil, Michael A. and Tu, Thomas},
-	month = nov,
-	year = {2017},
-	keywords = {Energy efficiency, Appliances, Data mining, Web crawling},
-	pages = {453--464},
-}
-
-@article{Blair:2014,
[…]
