+++ docs/source/references.bib
@@ -1,0 +1,140 @@
+@misc{Ashrae:2013,
+	title = {ASHRAE 93:2003 - {Method} {Of} {Testing} {To} {Determine} {The} {Thermal} {Performance} {Of} {Solar} {Collectors}},
+    author = {ASHRAE},
+	publisher = {American Society of Heating, Refrigerating \& Air Conditioning Engineer},
+	year = {2013},
+}
+
+@article{Blair:2014,
+	title = {System {Advisor} {Model}, {SAM} 2014.1.14: {General} {Description}},
+	language = {en},
+	journal = {National Renewable Energy Lab. (NREL), Golden, CO},
+	author = {Blair, Nate and Dobos, Aron P and Freeman, Janine and Neises, Ty and Wagner, Michael and Ferguson, Tom and Gilman, Paul and Janzou, Steven},
+	year = {2014},
+	pages = {19},
+}
+
+@misc{Cec_appliance:2019,
+	title = {Appliance {Certification} {Database}},
+	url = {https://cacertappliances.energy.ca.gov/},
+	author = {CEC},
+	year = {2019},
+}
+
+@misc{Ccms:2018,
+	title = {Compliance {Certification} {Management} {System}},
+	url = {https://www.regulations.doe.gov/ccms},
+	urldate = {2018-12-12},
+	author = {{U.S. Department of Energy - Appliance \& Equipment Standards Program}},
+    year = {2018},
+}
+
+@techreport{Coughlin:2020,
+	title = {Costs and {Benefits} of {Community} versus {Individual} {End}-use {Infrastructure} for {Solar} {Water} {Heating}},
+	institution = {Lawrence Berkeley National Laboratory},
+	author = {Coughlin, Katie and Grahovac, Milica and Ganeshalingam, Mohan and Hosbach, Robert and Vossos, Vagelis},
+	year = {2021},
+	publisher = {California Energy Commission},
+}
+
+@techreport{Csi_thermal:2016,
+	title = {California {Solar} {Initiative} - {Thermal} {Program} {Handbook}},
+	language = {en},
+	journal = {CPUC},
+	author = {CPUC},
+	year = {2016},
+	pages = {179},
+}
+
+@manual{DiOrio:2014,
+	title = {Technical Manual for the SAM Solar Water Heating Model},
+	language = {en},
+	journal = {National Renewable Energy Lab. (NREL), Golden, CO},
+	author = {DiOrio, Nicholas and Christensen, Craig and Burch, Jay and Dobos, Aron},
+	year = {2014},
+	url = {https://sam.nrel.gov/solar-water-heating.html},
+	pages = {24},
+}
+
+@book{Duffie:2013,
+	title = {Solar {Engineering} of {Thermal} {Processes}},
+	language = {en},
+	author = {Dufﬁe, John A and Beckman, William A},
+	pages = {928},
+    publisher = {John Wiley & Sons, Inc., Hoboken, New Jersey},
+    isbn = {978-0-470-87366-3},
+    year = {2013},
+}
+
+@article{Gerke:2017,
+	title = {The {International} {Database} of {Efficient} {Appliances} ({IDEA}): {A} new tool to support appliance energy-efficiency deployment},
+	volume = {205},
+	issn = {0306-2619},
+	shorttitle = {The {International} {Database} of {Efficient} {Appliances} ({IDEA})},
+	url = {http://www.sciencedirect.com/science/article/pii/S0306261917309650},
+	doi = {10.1016/j.apenergy.2017.07.093},
+	journal = {Applied Energy},
+	author = {Gerke, Brian F. and McNeil, Michael A. and Tu, Thomas},
+	month = nov,
+	year = {2017},
[…]
