+++ docs/source/bibliography.rst
@@ -1,5 +1,3 @@
-.. _References:
-
 References
 ==========
 
