+++ docs/source/acknowledgements.rst
@@ -1,0 +1,4 @@
+Acknowledgements
+================
+
+This work was supported by the California Energy Commission, Public Interest Energy Research Program, under Contract No. PIR-16-022.
