+++ docs/source/references.bib
@@ -1,3 +1,5 @@
+% Encoding: UTF-8
+
 @misc{Ashrae:2013,
 	title = {ASHRAE 93:2003 - {Method} {Of} {Testing} {To} {Determine} {The} {Thermal} {Performance} {Of} {Solar} {Collectors}},
     author = {ASHRAE},
