### 2020-12-02 f1320e3
was able to generate docs with references
