+++ paper/paper.bib
@@ -33,21 +33,3 @@
     year = {2020},
 	url = {https://aceee2020.conferencespot.org/event-data/pdf/catalyst_activity_10923/catalyst_activity_paper_20200812133157248_498ce455_3a9c_4278_9088_6e3fdce5745b},
 	}
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
