+++ docs/source/references.bib
@@ -31,7 +31,7 @@
     year = {2018},
 }
 
-@techreport{Coughlin:2020,
+@techreport{Coughlin:2021,
 	title = {Costs and {Benefits} of {Community} versus {Individual} {End}-use {Infrastructure} for {Solar} {Water} {Heating}},
 	institution = {Lawrence Berkeley National Laboratory},
 	author = {Coughlin, Katie and Grahovac, Milica and Ganeshalingam, Mohan and Hosbach, Robert and Vossos, Vagelis},
@@ -42,6 +42,7 @@
 @techreport{Csi_thermal:2016,
 	title = {California {Solar} {Initiative} - {Thermal} {Program} {Handbook}},
 	language = {en},
+	institution = {California Energy Commission},
 	journal = {CPUC},
 	author = {CPUC},
 	year = {2016},
@@ -83,7 +84,7 @@
 	pages = {453--464},
 }
 
-@thesis{Gerhart:2019,
+@misc{Gerhart:2019,
 	title = {Implementation of a {Flexible} {Web} {Framework} for {Simulating} {Python} {System} {Models}},
 	language = {en},
 	author = {Gerhart, Hannes},
