+++ README.md
@@ -126,3 +126,5 @@ This work was supported by the California Energy Commission, Public Interest Ene
 
 We thank the reviewers and the editor of [The Journal of Open Source Software (JOSS)](https://joss.theoj.org/), [Bryn Pickering](https://github.com/brynpickering), [Nithiya Streethran](https://github.com/nmstreethran), and [Stefan Pfenninger](https://github.com/sjpfenninger) for their contributions in improving the code, the examples and the code documentation for the code release 2.0.0.
 
+*mg add link to pdf of the compiled documentation
+
