+++ paper/paper.bib
@@ -1,8 +1,8 @@
 @techreport{Coughlin:2020,
 	title = {Costs and {Benefits} of {Community} versus {Individual} {End}-use {Infrastructure} for {Solar} {Water} {Heating}},
-	institution = {Lawrence Berkeley National Laboratory},
+	institution = {California Energy Commission},
 	author = {Coughlin, Katie and Grahovac, Milica and Ganeshalingam, Mohan and Hosbach, Robert and Vossos, Vagelis},
-	year = {2021},
+	year = {2020},
 	publisher = {California Energy Commission},
 }
 
@@ -15,12 +15,13 @@
 	month = {may},
 }
 
-@thesis{Gerhart:2019,
+@masterthesis{Gerhart:2019,
 	title = {Implementation of a {Flexible} {Web} {Framework} for {Simulating} {Python} {System} {Models}},
 	language = {en},
 	author = {Gerhart, Hannes},
 	pages = {82},
     year = {2019},
+    institution = {Technical University of Munich},
     publisher = {Technical University of Munich},
 }
 
