+++ docs/source/references.bib
@@ -10,7 +10,7 @@
 @article{Blair:2014,
 	title = {System {Advisor} {Model}, {SAM} 2014.1.14: {General} {Description}},
 	language = {en},
-	journal = {National Renewable Energy Lab. (NREL), Golden, CO},
+	journal = {National Renewable Energy Laboratory (NREL), Golden, CO},
 	author = {Blair, Nate and Dobos, Aron P and Freeman, Janine and Neises, Ty and Wagner, Michael and Ferguson, Tom and Gilman, Paul and Janzou, Steven},
 	year = {2014},
 	pages = {19},
@@ -33,7 +33,7 @@
 
 @techreport{Coughlin:2021,
 	title = {Costs and {Benefits} of {Community} versus {Individual} {End}-use {Infrastructure} for {Solar} {Water} {Heating}},
-	institution = {Lawrence Berkeley National Laboratory},
+	institution = {California Energy Commission},
 	author = {Coughlin, Katie and Grahovac, Milica and Ganeshalingam, Mohan and Hosbach, Robert and Vossos, Vagelis},
 	year = {2021},
 	publisher = {California Energy Commission},
@@ -90,6 +90,7 @@
 	author = {Gerhart, Hannes},
 	pages = {82},
     year = {2019},
+    institution = {Technical University of Munich},
     publisher = {Technical University of Munich},
 }
 
