+++ paper/paper.bib
@@ -1,8 +1,8 @@
-@techreport{Coughlin:2020,
+@techreport{Coughlin:2021,
 	title = {Costs and {Benefits} of {Community} versus {Individual} {End}-use {Infrastructure} for {Solar} {Water} {Heating}},
-	institution = {California Energy Commission},
+	institution = {Lawrence Berkeley National Laboratory},
 	author = {Coughlin, Katie and Grahovac, Milica and Ganeshalingam, Mohan and Hosbach, Robert and Vossos, Vagelis},
-	year = {2020},
+	year = {2021},
 	publisher = {California Energy Commission},
 }
 
