+++ docs/source/references.bib
@@ -33,7 +33,7 @@
 
 @techreport{Coughlin:2021,
 	title = {Costs and {Benefits} of {Community} versus {Individual} {End}-use {Infrastructure} for {Solar} {Water} {Heating}},
-	institution = {California Energy Commission},
+	institution = {Lawrence Berkeley National Laboratory},
 	author = {Coughlin, Katie and Grahovac, Milica and Ganeshalingam, Mohan and Hosbach, Robert and Vossos, Vagelis},
 	year = {2021},
 	publisher = {California Energy Commission},
