### 2020-12-02 40b2a73
removed duplicate text, checked spelling, checked word count
