+++ docs/source/references.bib
@@ -33,7 +33,8 @@
 
 @techreport{Coughlin:2021,
 	title = {Costs and {Benefits} of {Community} versus {Individual} {End}-use {Infrastructure} for {Solar} {Water} {Heating}},
-	institution = {Lawrence Berkeley National Laboratory},
+	affiliation = {Lawrence Berkeley National Laboratory},
+    institution = {California Energy Commission},
 	author = {Coughlin, Katie and Grahovac, Milica and Ganeshalingam, Mohan and Hosbach, Robert and Vossos, Vagelis},
 	year = {2021},
 	publisher = {California Energy Commission},
