### 2020-12-02 75e1792
moved text sections, changed bibliography style
