+++ _sources/source/models.rst.txt
@@ -1,47 +1,112 @@
-.. _sec_sys_mod:
-
 Multiscale Solar Water Heating (MSWH)
--------------------------------------
+=====================================
 
 Scope
 ^^^^^
 
 The main purpose of the Multiscale Solar Water Heating (MSWH) software is to model energy use for individual and community scale solar water heating projects in California.
 
-The package contains functional and unit tests and it is structured so that it can be extended with further technologies, applications and locations.
+The package contains functional and unit tests and it is structured so that it can be extended with further technologies, applications, and locations.
 
 Usage
 ^^^^^
 
-The user provides a climate zone for a project, an occupancy for each household and whether any of the occupants stay at home during the day. The software can then load a set of example California specific hourly domestic hot water end-use load profiles from a database, size and locate the systems. The user can now simulate the hourly system performance over a period of one representative year, visualize and explore the simulation results using time-series plots for temperature profiles, heat and power rates, or look at annual summaries. Similarly the user can model individual household solar water heating projects and base case conventional gas tank water heater systems, such that the results can be compared between the individual, community and base case systems.
+The user provides a climate zone for a project, occupancy for each household, and whether any of the occupants stay at home during the day. The software can then load a set of example California specific hourly domestic hot water end-use load profiles from a database, size, and locate the systems. Next, the user can simulate the hourly system performance over a period of one representative year, visualize and explore the simulation results using time-series plots for temperature profiles, heat and power rates, or look at annual summaries. Similarly, the user can model individual household solar water heating projects and base case conventional gas tank water heater systems, such that the results can be compared between the individual, community-scale, and base case systems.
 
-This functionality is readily available through a Jupyter notebook and a Django web framework, depending on what level of detail the user would like to access. Please see the README file on the `MSWH repo <https://github.com/LBNL-ETA/MSWH>`_ for usage and installation details.
+This functionality is readily available through a `Jupyter notebook <https://github.com/LBNL-ETA/MSWH/blob/v2.0.0/scripts/MSWH%20System%20Tool.ipynb>`_ and a `Django web framework graphical user interface (GUI) <https://github.com/LBNL-ETA/MSWH/tree/v2.0.0/web>`_, depending on what level of detail the user would like to access. Please see the README file on the `MSWH repo <https://github.com/LBNL-ETA/MSWH>`_ for further usage and installation instructions.
+
+System performance time series visualizations are available both in example notebooks and through the GUI, either spun off locally or `using a web deployed version <https://solar.floweragenda.org/>`_.
 
 Features
 ^^^^^^^^
 
 This software package contains the following Python modules:
 
-* Solar irradiation on a tilted surface
+* Solar irradiation on a tilted surface.
 
-* Simplified component models for Converter (solar collectors, electric resistance heater, gas burner, photovoltaic panels, heat pump), Storage (solar thermal tank, heat pump thermal tank, conventional gas tank water heater), and Distribution (distribution and solar pump, piping losses) components
+* Simplified component models for:
 
-* Preconfigured system simulation models for: base case gas tank water heaters, solar thermal water heaters (solar collector feeding a storage tank, with a tankless gas water heater backup in a new installation cases and a basecase gas tank water heater in a retrofit case) and solar electric water heaters (heat pump storage tank with an electric resistance backup)
+    * Converters: solar collectors, electric resistance heater, gas burner, photovoltaic panels, heat pump.
+    * Storage: solar thermal tank, heat pump thermal tank, conventional gas tank water heater.
+    * Distribution: distribution and solar pump, piping losses.
 
-* Database with component performance parameters, California specific weather data and domestic hot water end-use load profiles
+* Preconfigured system simulation models for: 
 
[…]
