+++ _sources/source/bibliography.rst.txt
@@ -1,0 +1,6 @@
+References
+==========
+
+.. bibliography:: references.bib
+   :all:
+   :style: alpha
