+++ _sources/index.rst.txt
@@ -1,8 +1,3 @@
-.. MSWH documentation master file, created by
-   sphinx-quickstart on Thu Sep 13 15:36:37 2018.
-   You can adapt this file completely to your liking, but it should at least
-   contain the root `toctree` directive.
-
 Multiscale Solar Water Heating System Modeling
 ==============================================
 
@@ -14,6 +9,8 @@ Multiscale Solar Water Heating System Modeling
    source/installation
    source/models
    source/modules
+   source/acknowledgements
+   source/bibliography
 
    source/legal
 
