### 2020-12-02 #2695 @milicag
@sjpfenninger I moved the parts of the paper, as you suggested, into the documentation. More specifically, the portions of the statement of need now are hosted [here](https://lbnl-eta.github.io/MSWH/source/models.html#future-applications-statement-of-need), while the parts related to the approach and methodology can be found [here](https://lbnl-eta.github.io/MSWH/source/models.html#approach-to-component-and-system-modeling-and-simulation). The paper is now significantly shorter. I summarized the parts that were extracted from the paper where I thought it was suitable.

Please let me know if you need any further edits. I can retag once we have the final version.
