+++ paper/paper.bib
@@ -10,6 +10,7 @@
 	title = {Multiscale Solar Water Heating (MSWH) v1.0},
 	author = {Grahovac, Milica and Hosbach, Robert and Coughlin, Katie and Ganeshalingam, Mohan and Gerhart, Hannes},
 	url = {https://doi.org/10.11578/dc.20190524.2},
+	doi = {10.11578/dc.20190524.2},
 	howpublished = {[Computer Software] \url{https://doi.org/10.11578/dc.20190524.2}},
 	year = {2019},
 	month = {may},
