### 2020-12-03 #2695 @sjpfenninger
@milicag I made some minor suggestions for edits in https://github.com/LBNL-ETA/MSWH/pull/39. One additional minor point: you refer to a "12-core processor" and run time on that system, can you specify what type of processor to make this a more useful benchmark?
