### 2020-12-03 e49c5ab
added processor details
