### 2020-12-03 #2695 @milicag
@sjpfenninger Thanks for your edits. I merged the PR. I specified the processor by just copying the info from my system. Please let me know if there is anything else that needs to be done before the re-tag.
