### 2020-12-04 #2695 @sjpfenninger
@milicag Great, thanks. So yes, we need you to make a tagged release on GitHub and archive that release, for example on Zenodo or figshare, then report the version number and archive DOI back here in the review issue. Make sure the archive has the same author list and title as the paper does.
