### 2020-12-04 #2695 @milicag
@sjpfenninger, here are the DOI of the v2 archive at Zenodo and a new release version tag:

[The tagged release](https://github.com/LBNL-ETA/MSWH/releases/tag/v2.0.0) with a version number v2.0.0
[DOI: 10.5281/zenodo.4305504](https://zenodo.org/record/4305504)
