### 2020-12-04 #2695 @sjpfenninger
Hm, the Zenodo deposition says "Closed Access -Files are not publicly accessible."

Not sure what went wrong there, but can you fix that?
