### 2020-12-04 #2695 @milicag
@sjpfenninger Zenodo was asking me to associate the code archive with an another license, I thought it was much safer to provide a link to our repository as an archive that contains the zip files ready for download of the version 2.0.0 of the software including the paper. The users can use the links provided in the Zenodo archive description to download the release from the Github repository. Is there an issue with this approach?
