### 2020-12-04 #2695 @milicag
@sjpfenninger This is the [initial link on the Zenodo](https://github.com/LBNL-ETA/MSWH/releases/tag/v2.0.0) summary. This way I preserve only one layer of licensing that was granted to us through DOE code publication.
