### 2020-12-04 #2695 @milicag
@sjpfenninger I could change the link to be [a direct download of the release v2.0.0](https://github.com/LBNL-ETA/MSWH/archive/v2.0.0.zip), or potentially better I could add that link in addition to the above link.
