### 2020-12-04 #2695 @milicag
With my approach I avoided a need to use an additional license, while obtaining a DOI for the archive and providing full access to the code under a JOSS accepted license.
