### 2020-12-04 #2695 @milicag
@brynpickering Just to avoid any potential confusion, my knowledge in the domain of license interaction, especially about aspects such as which license takes precedence and where and how that is exactly defined, and whether it might change in the future, is limited. Therefore I found a simplified solution that I don't see imposes any limitations provided the information available at the JOSS website. I am certainly open to hearing your view, in particular if there is something within the JOSS rules that would conflict with the approach I've taken. Please do not hesitate to ask any questions you might have, I am happy to learn more.
