### 2020-12-07 #2695 @sjpfenninger
@milicag We are not asking you to chose a different license for the Zenodo archive. I do see that the LBNL BSD license is not a pre-defined choice on Zenodo. However you can choose as license "Other (Open)", then ensure that your license file which contains the BSD LBNL license is included at the root level of the Zenodo deposition (this should be no issue, as it's in the root of your repository).

We do need the archive on Zenodo to be a full archive of the submitted code, not just an empty submission to get a DOI. Let me know if this clears things up!
