### 2020-12-07 #2695 @sjpfenninger
@milicag: Great. The license is set to "Creative Commons Attribution 4.0 International" though. Can you set it to "Other (Open)"?
