### 2020-12-07 #2695 @sjpfenninger
@milicag Actually I checked with Zenodo support and the license should be available However it is called "Lawrence Berkeley National Labs BSD variant license" on Zenodo (if you enter some or all of those terms in the search field on upload, you should see it). If you can pick that license, the Zenodo deposition gets the correct license metadata and all should be well.
