+++ web/swhweb/wsgi.py
@@ -8,7 +8,21 @@ https://docs.djangoproject.com/en/2.1/howto/deployment/wsgi/
 """
 
 import os
+import sys
 
+# import paths from settings.py
+from .local_settings import DJANGO_PROJECT_PATH, VIRTUALENV_PACKAGES_PATH, MSWH_PROJECT_PATH
+
+# add the path to this file
+sys.path.append(DJANGO_PROJECT_PATH)
+
+# add the virtualenv site-packages path
+sys.path.append(VIRTUALENV_PACKAGES_PATH)
+
+# add the path to the mswh Python package
+sys.path.append(MSWH_PROJECT_PATH)
+
+# import Django wsgi implementation only after appending all paths to SYS paths
 from django.core.wsgi import get_wsgi_application
 
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "swhweb.settings")
