### 2020-12-07 aacd59c
added template for local settings
