### 2020-12-07 11765f1
added some instructions for deployment
