+++ web/swhweb/settings.py
@@ -145,6 +145,9 @@ STATICFILES_DIRS = [
     ),  # statics not tied to any specific app
 ]
 
+# Domain of landing page if web app is deployed on subdomain. Override this in local_settings.py
+HOME_DOMAIN = None
+
 # Read information from local_settings.py and overwrite values specified in this file
 # Keep these lines at the end of the file
 try:
