### 2020-12-07 288d718
add home domain to local settings template
