### 2020-12-07 a652fe7
moved home button text into settings
