### 2020-12-07 2bfba33
moved home button text into settings
