### 2020-12-08 0bc54d2
Added license section
