### 2020-12-08 #2695 @milicag
@sjpfenninger, this was an interesting and new journey! Thank you all for the contibutions and improvements to the documentation, the code, and the paper. I learned a lot.
