### 2020-12-14 #2695 @arfon
@brynpickering, @nmstreethran - many thanks for your reviews here and to @sjpfenninger for editing this submission. JOSS relies upon the volunteer efforts of people like yourselves, we simply wouldn't be able to do it without you! :sparkles:
